feat: v0.3.0 honest-engine — extract runtime, cost estimates, structural eval
Move mocks, RAPTOR, and leftover LangGraph nodes out of app.py; unify GraphState; add estimate/eval CLI; document metaphors vs real ops.
