+++ .github/workflows/ci.yml
@@ -39,22 +39,27 @@ jobs:
           ruff check
           deepthink/config.py deepthink/cli.py deepthink/__main__.py
           deepthink/providers.py deepthink/api.py deepthink/__init__.py
-          deepthink/distillation
-          tests/run_all.py tests/phase_config.py tests/phase_package_api.py tests/_paths.py
+          deepthink/cost.py deepthink/sessions.py deepthink/eval_structural.py
+          deepthink/runtime/bus.py deepthink/distillation
+          tests/run_all.py tests/phase_config.py tests/phase_package_api.py
+          tests/phase_honesty.py tests/test_control_flow.py tests/_paths.py
 
       - name: Ruff format (check only)
         run: >
           ruff format --check
           deepthink/config.py deepthink/cli.py deepthink/__main__.py
           deepthink/providers.py deepthink/api.py deepthink/__init__.py
-          deepthink/distillation
-          tests/_paths.py tests/run_all.py tests/phase_config.py tests/phase_package_api.py
+          deepthink/cost.py deepthink/sessions.py deepthink/eval_structural.py
+          deepthink/runtime/bus.py deepthink/distillation
+          tests/_paths.py tests/run_all.py tests/phase_config.py
+          tests/phase_package_api.py tests/phase_honesty.py tests/test_control_flow.py
 
       - name: Mypy (typed surface)
         run: >
           mypy --config-file pyproject.toml
           deepthink/config.py deepthink/cli.py deepthink/__main__.py
           deepthink/providers.py deepthink/api.py
+          deepthink/cost.py deepthink/sessions.py deepthink/runtime/bus.py
 
       - name: Ensure env template present for static tests
         run: |
