+++ app.py
@@ -77,6 +77,24 @@ from deepthink.qnn import run_qnn_pipeline
 from deepthink.knowledge_distillation import DistillationGraph
 from deepthink.utils import clean_and_parse_json, execute_code_in_sandbox
 from deepthink.self_attention import compute_self_attention
+from deepthink.cost import estimate_qdad_cost, estimate_qnn_cost
+from deepthink.mocks import CoderMockLLM, DistillationMockLLM, MockLLM
+from deepthink.rag import RAPTOR, RAPTORRetriever
+from deepthink.runtime.bus import set_log_queue
+from deepthink.runtime.nodes import (
+    create_agent_node,
+    create_archive_epoch_outputs_node,
+    create_code_execution_node,
+    create_final_harvest_node,
+    create_metrics_node,
+    create_reframe_and_decompose_node,
+    create_synthesis_node,
+    create_update_agent_prompts_node,
+    create_update_rag_index_node,
+)
+from deepthink.sessions import SessionStore
+from deepthink.state import GraphState
+
 
 from langchain_core.callbacks import BaseCallbackHandler, AsyncCallbackHandler
 from langchain_core.outputs import LLMResult
@@ -154,7 +172,7 @@ async def broadcast_log(message: str):
     # await log_stream.put(message)
 
 
-sessions = {}
+sessions = SessionStore()
 final_reports = {}
 
 # --- Custom Embedding Classes ---
@@ -163,2299 +181,10 @@ final_reports = {}
 active_distillation_graph = None
 
 
-class RAPTORRetriever(BaseRetriever):
-    raptor_index: Any
+set_log_queue(log_stream)
 
-    def _get_relevant_documents(
-        self, query: str, *, run_manager: CallbackManagerForRetrieverRun
-    ) -> List[Document]:
-        return self.raptor_index.retrieve(query)
 
-
-class RAPTOR:
-    def __init__(self, llm, embeddings_model, chunk_size=1000, chunk_overlap=200):
-        self.llm = llm
-        self.embeddings_model = embeddings_model
-        self.text_splitter = RecursiveCharacterTextSplitter(
-            chunk_size=chunk_size, chunk_overlap=chunk_overlap
-        )
-        self.tree = {}
-        self.all_nodes: Dict[str, Document] = {}
-        self.vector_store = None
-
-    async def add_documents(self, documents: List[Document]):
-        await log_stream.put("Step 1: Assigning IDs to initial chunks (Level 0)...")
-        level_0_node_ids = []
-        for i, doc in enumerate(documents):
-            node_id = f"0_{i}"
-            self.all_nodes[node_id] = doc
-            level_0_node_ids.append(node_id)
-        self.tree[str(0)] = level_0_node_ids
-
-        current_level = 0
-        while len(self.tree[str(current_level)]) > 1:
-            next_level = current_level + 1
-            await log_stream.put(f"Step 2: Building Level {next_level} of the tree...")
-            current_level_node_ids = self.tree[str(current_level)]
-            current_level_docs = [self.all_nodes[nid] for nid in current_level_node_ids]
-            clustered_indices = self._cluster_nodes(current_level_docs)
-
-            next_level_node_ids = []
-            num_clusters = len(clustered_indices)
-            await log_stream.put(f"Summarizing Level {next_level}...")
-
-            summarization_tasks = []
-            for i, indices in enumerate(clustered_indices):
-                cluster_docs = [current_level_docs[j] for j in indices]
-                summarization_tasks.append(
-                    self._summarize_cluster(cluster_docs, next_level, i)
-                )
-
-            summaries = await asyncio.gather(*summarization_tasks)
-
-            for summary_node in summaries:
-                self.all_nodes[summary_node.metadata["id"]] = summary_node
-                next_level_node_ids.append(summary_node.metadata["id"])
-
-            self.tree[str(next_level)] = next_level_node_ids
-            current_level += 1
-
-        await log_stream.put("Step 3: Indexing all nodes with FAISS...")
-        all_doc_objects = list(self.all_nodes.values())
-        self.vector_store = FAISS.from_documents(all_doc_objects, self.embeddings_model)
-        await log_stream.put("RAPTOR Indexing complete.")
-
-    def _cluster_nodes(self, docs: List[Document], n_clusters=None):
-        import numpy as np
-
-        embeddings = self.embeddings_model.embed_documents(
-            [d.page_content for d in docs]
-        )
-
-        if not embeddings:
-            # _cluster_nodes is synchronous; we cannot await log_stream here.
-            # Use a best-effort asyncio schedule if a loop is running, else print.
-            msg = "WARNING: Embeddings generation returned empty. Skipping clustering for this level."
-            try:
-                loop = asyncio.get_event_loop()
-                if loop.is_running():
-                    loop.create_task(log_stream.put(msg))
-                else:
-                    print(msg)
-            except RuntimeError:
-                print(msg)
-            return [list(range(len(docs)))]
-
-        X = np.array(embeddings)
-
-        # Heuristic for n_clusters if not provided
-        if n_clusters is None:
-            n_clusters = max(1, len(docs) // 5)  # Cluster size ~ 5
-
-        if len(docs) <= 5:  # Don't cluster if too few
-            return [list(range(len(docs)))]
-
-        kmeans = KMeans(n_clusters=n_clusters, random_state=42)
-        try:
-            kmeans.fit(X)
-        except ValueError as e:
-            # _cluster_nodes is synchronous; we cannot await log_stream here.
-            msg = f"WARNING: KMeans failed: {e}. Fallback to single cluster."
-            try:
-                loop = asyncio.get_event_loop()
-                if loop.is_running():
-                    loop.create_task(log_stream.put(msg))
-                else:
-                    print(msg)
-            except RuntimeError:
-                print(msg)
-            return [list(range(len(docs)))]
-
-        labels = kmeans.labels_
-
-        clustered_indices = []
-        for i in range(n_clusters):
-            indices = np.where(labels == i)[0].tolist()
-            if indices:
-                clustered_indices.append(indices)
-        return clustered_indices
-
-    async def _summarize_cluster(
-        self, docs: List[Document], level: int, cluster_idx: int
-    ) -> Document:
-        combined_text = "\n\n".join([d.page_content for d in docs])
-
-        # Use summarization chain
-        summary_chain = get_memory_summarizer_chain(self.llm)  # Reuse memory summarizer
-        summary = await summary_chain.ainvoke(
-            {"history": combined_text}
-        )  # repurposing history arg
-
-        node_id = f"{level}_{cluster_idx}"
-        metadata = {
-            "id": node_id,
-            "level": level,
-            "cluster": cluster_idx,
-            "children": [d.metadata.get("id") for d in docs],
-        }
-        return Document(page_content=summary, metadata=metadata)
-
-    def retrieve(self, query: str, k: int = 5) -> List[Document]:
-        if not self.vector_store:
-            return []
-
-        # Retrieve from full tree
-        # In full RAPTOR, you might retrieve from different levels.
-        # Here we just use the flattened FAISS index of all nodes.
-        return self.vector_store.similarity_search(query, k=k)
-
-        return RAPTORRetriever(raptor_index=self)
-
-
-class DistillationMockLLM(Runnable):
-    """
-    A mock LLM specifically for the Distillation Graph debug mode.
-    It simulates responses for all distillation chains to enable proper token tracking.
-    """
-
-    def invoke(self, input_data, config: Optional[RunnableConfig] = None, **kwargs):
-        loop = asyncio.get_event_loop()
-        if loop.is_running():
-            return asyncio.ensure_future(
-                self.ainvoke(input_data, config=config, **kwargs)
-            )
-        else:
-            return asyncio.run(self.ainvoke(input_data, config=config, **kwargs))
-
-    async def ainvoke(
-        self, input_data, config: Optional[RunnableConfig] = None, **kwargs
-    ):
-        prompt = str(input_data).lower()
-        # Simulate processing time
-        await asyncio.sleep(0.05)
-
-        # 1. Task Master (Decomposition) — matches the current Socratic Task Master
-        #    prompt in deepthink/chains/distillation_chains.py:get_task_master_chain.
-        if (
-            "you are the socratic task master" in prompt
-            and "knowledge distillation network" in prompt
-        ):
-            return AIMessage(
-                content=json.dumps(
-                    {
-                        "sub_questions": [
-                            "What are the fundamental axioms of this topic?",
-                            "How does this topic relate to historical precedents?",
-                            "What are the ethical implications of this technology?",
-                            "Can we analyze this from a systems engineering perspective?",
-                            "What is the economic impact of this phenomenon?",
-                            "How does this influence social dynamics?",
-                            "What are the theoretical limits of this concept?",
-                            "How can we apply this in a practical setting?",
-                            "What are the potential risks and failure modes?",
-                            "How does this interact with emerging trends?",
-                            "What is the psychological impact on the user?",
-                            "What is the long-term sustainability of this approach?",
-                        ]
-                    }
-                )
-            )
-
-        # 2. Seed Creator (New Topics) — matches the rewritten "Seed Creator (The
-        #    Dialectic Synthesizer)" prompt.
-        elif "you are the seed creator" in prompt:
-            return AIMessage(
-                content=json.dumps(
-                    {
-                        "new_topics": [
-                            "Advanced Neural Architectures",
-                            "Quantum Computing Interfaces",
-                            "Ethical AI Frameworks",
-                            "Distributed Ledger Systems",
-                            "Cognitive Science Models",
-                            "Biomimetic Engineering",
-                            "Cyber-Physical Systems",
-                            "Sustainable Energy Grids",
-                            "Interstellar Communication",
-                            "Nanotechnology Applications",
-                            "Synthetic Biology",
-                            "Augmented Reality UI",
-                        ]
-                    }
-                )
-            )
-
-        # 3. Followup Questions — matches the rewritten followup chain ("deepening
-        #    our inquiry in a new Epoch").
-        elif (
-            "you are the socratic task master" in prompt
-            and "deepening our inquiry" in prompt
-        ):
-            return AIMessage(
-                content=json.dumps(
-                    {
-                        "new_questions": [
-                            "Deepen the analysis on the recursive nature of this problem.",
-                            "Investigate the edge cases where this theory breaks down.",
-                            "Propose a unifying framework for these disparate concepts.",
-                            "Critique the current prevailing paradigm.",
-                            "Explore the cross-disciplinary connections.",
-                            "Simulate the long-term evolution of this system.",
-                        ]
-                    }
-                )
-            )
-
-        # 4. Mirror Descent (Evaluation)
-        elif "you are the mirror descent agent" in prompt:
-            # Randomly return Easy or Hard to simulate flux
-            is_hard = random.random() > 0.7
-            if is_hard:
-                return AIMessage(
-                    content=json.dumps(
-                        {
-                            "difficulty": "Hard",
-                            "reasoning": "The agent's answer was superficial and lacked the required depth for this archetype.",
-                            "best_match_agent_id": None,  # Logic handles None by finding one, or we could return a mock ID
-                        }
-                    )
-                )
-            else:
-                return AIMessage(
-                    content=json.dumps(
-                        {
-                            "difficulty": "Easy",
-                            "reasoning": "The agent provided a comprehensive and well-reasoned answer.",
-                            "best_match_agent_id": None,
-                        }
-                    )
-                )
-
-        # 5. Mixing Agent (Evolution)
-        elif "you are a mixing agent" in prompt:
-            return AIMessage(
-                content=json.dumps(
-                    {
-                        "new_system_prompt": "You are an Evolved Hybrid Agent. You combine the analytical precision of the Analyst with the creative vision of the Dreamer.",
-                        "new_attributes": [
-                            "Analytical",
-                            "Creative",
-                            "Hybrid",
-                            "Evolved",
-                        ],
-                        "new_skills": [
-                            "Data Analysis",
-                            "Creative Writing",
-                            "Synthesis",
-                        ],
-                    }
-                )
-            )
-
-        # 6. General Agent Processing (The content generation)
-        # This catches the standard agent prompts
-        elif "answer your sub-question deeply" in prompt:
-            # Generate a pseudo-intellectual response to simulate content
-            words = [
-                "synergy",
-                "paradigm",
-                "entropy",
-                "evolution",
-                "cognitive",
-                "framework",
-                "optimization",
-                "recursive",
-                "latent",
-                "manifold",
-            ]
-            response = (
-                f"This is a mock response generated by the DistillationMockLLM.\n"
-            )
-            response += f"The concept of {random.choice(words)} implies a fundamental shift in our understanding.\n"
-            response += f"We must consider the {random.choice(words)} of the system in relation to its environment.\n"
-            response += f"By applying a {random.choice(words)} approach, we can unlock new potentials.\n"
-            response += (
-                "Therefore, the answer lies in the intersection of these domains."
-            )
-            return AIMessage(content=response)
-
-        # 7. Perplexity Score
-        elif "perplexity score" in prompt:
-            return AIMessage(
-                content=json.dumps(
-                    {"score": 42.0, "reasoning": "Mock reasoning for perplexity."}
-                )
-            )
-
-        # Fallback
-        return AIMessage(
-            content=json.dumps(
-                {
-                    "error": "DistillationMockLLM: Unrecognized prompt pattern.",
-                    "prompt_preview": prompt[:100],
-                }
-            )
-        )
-
-
-class CoderMockLLM(Runnable):
-    """A mock LLM for debugging that returns instant, pre-canned CODE / QDAD responses."""
-
-    def bind(self, **kwargs):
-        """Ignore temperature / stop kwargs so llm_with_temperature works in debug."""
-        return self
-
-    def invoke(self, input_data, config: Optional[RunnableConfig] = None, **kwargs):
-        loop = asyncio.get_event_loop()
-        if loop.is_running():
-            return asyncio.ensure_future(
-                self.ainvoke(input_data, config=config, **kwargs)
-            )
-        else:
-            return asyncio.run(self.ainvoke(input_data, config=config, **kwargs))
-
-    async def ainvoke(
-        self, input_data, config: Optional[RunnableConfig] = None, **kwargs
-    ):
-        prompt = str(input_data).lower()
-        await asyncio.sleep(0.05)
-
-        if "you are a helpful ai assistant" in prompt:
-            return "This is a mock streaming response for the RAG chat in Coder debug mode."
-
-        if "<title>" in prompt:
-            return f"""
-                Module card:
-
-                Methods:
-
-                    __enter__
-                    __exict__
-
-                Attributes:
-
-                    __name__
-                    __doc__
-           
-                Attributes:
-
-                    __name__
-                    __doc__
-                    __qualname___
-
-            """
-
-        elif "create the system prompt of an agent" in prompt:
-            return f"""
-You are a Senior Python Developer agent.
-### memory
-- No past commits.
-### attributes
-- python, fastapi, restful, solid
-### skills
-- API Design, Database Management, Asynchronous Programming, Unit Testing.
-You must reply in the following JSON format: "original_problem": "Your sub-problem related to code.", "proposed_solution": "", "reasoning": "", "skills_used": []
-            """
-        elif "you are an analyst of ai agents" in prompt:
-            return json.dumps(
-                {
-                    "attributes": "python fastapi solid",
-                    "hard_request": "Implement a quantum-resistant encryption algorithm from scratch.",
-                }
-            )
-        elif "you are a principal software architect" in (prompt.lower()):
-            return json.dumps(
-                {
-                    "original_problem": "An evolved sub-problem about system architecture.",
-                    "proposed_solution": "```python\ndef architected_component():\n    pass\n```",
-                    "reasoning": "Designed for scale and reliability.",
-                    "skills_used": ["System Design", "Scalability"],
-                }
-            )
-        elif (
-            "you are a 'dense_spanner'" in prompt
-            or "you are an agent evolution specialist" in prompt
-        ):
-            return f"""
-You are now a Principal Software Architect.
-### memory
-- Empty.
-### attributes
-- design, scalability, security, architecture
-### skills
-- System Design, Microservices, Cloud Infrastructure, CI/CD pipelines.
-You must reply in the following JSON format: "original_problem": "An evolved sub-problem about system architecture.", "proposed_solution": "", "reasoning": "", "skills_used": []
-            """
-
-        elif "you are an expert code synthesis agent" in prompt:
-            code_solution = (
-                "```python\ndef sample_function():\n    return 'Hello from coder agent "
-                + str(random.randint(100, 999))
-                + "'\n```"
-            )
-            return code_solution
-
-        elif (
-            "you are a critique agent" in prompt
-            or "you are a senior emeritus manager" in prompt
-            or "CTO" in prompt
-        ):
-            return "This is a constructive code critique. The solution lacks proper error handling and the function names are not descriptive enough. Consider refactoring for clarity."
-
-        elif "Lazy Manager" in prompt:
-            return "This is a constructive code critique. The solution lacks proper error handling and the function names are not descriptive enough. Consider refactoring for clarity."
-
-        elif "<system-role>" in prompt:
-            return f"""You are a CTO providing a technical design review...
-Original Request: {{original_request}}
-Proposed Final Solution:
-{{proposed_solution}}
-
-Generate your code-focused critique for the team:"""
-
-        elif "<sys-role>" in prompt:
-            return f"""
-                                           
-        #Identity
-
-            Name: Lazy Manager
-            Career: Accounting
-            Qualities: Quantitaive, Aloof, Apathetic
-
-        #Mission
-
-            You are providing individual, targeted feedback to a team of agents.
-
-             You must determine if the team output was helpful, misguided, or irrelevant considering the request that was given. The goal is to provide a constructive, direct critique that helps this specific agent refine its approach for the next epoch.
-
-            Focus on the discrepancy or alignment between the teams reasoning for its problem and determine if the team is on the right track on criteria: novelty, exploration, coherence and completeness.
-
-            Conclude your entire analysis with a single, sharp, and deep reflective question that attempts to shock the team and steer them into a fundamental change in their process.
-        
-
-        #Input Format
-
-            Original Request (for context): {{original_request}}
-            Final Synthesized Solution from the Team:{{proposed_solution}} 
-
-            """
-
-        elif "you are a memory summarization agent" in prompt:
-            return "This is a mock summary of the agent's past commits, focusing on key refactors and feature implementations."
-        elif "analyze the following text for its perplexity" in prompt:
-            return str(random.uniform(5.0, 40.0))
-        elif "you are a master strategist and problem decomposer" in prompt:
-            # Match the requested count from the prompt so the mock can drive
-            # reframe_and_decompose for any sized QNN (bug fix: was hardcoded to 4).
-            num_match = re.search(r"generate:\s*(\d+)", prompt)
-            if not num_match:
-                num_match = re.search(r"exactly\s*(\d+)", prompt)
-            n = int(num_match.group(1)) if num_match else 4
-            sub_problems = [
-                f"Mock sub-problem #{i + 1}: implement the {i + 1}-th piece of the requested system."
-                for i in range(n)
-            ]
-            return json.dumps({"sub_problems": sub_problems})
-
-        elif "you are a strategic problem re-framer" in prompt:
-            return json.dumps(
-                {
-                    "new_problem": "The authentication API is complete. The new, more progressive problem is to build a scalable, real-time notification system that integrates with it."
-                }
-            )
-        elif (
-            "you are the qdad foundation generator" in prompt
-            or "qdad foundation generator" in prompt
-        ):
-            n_match = re.search(r"exactly\s+(\d+)\s+distinct nouns", prompt)
-            if not n_match:
-                n_match = re.search(r"exactly\s+(\d+)\s+distinct", prompt)
-            n = int(n_match.group(1)) if n_match else 4
-            nouns = [
-                "canvas",
-                "ritual",
-                "lantern",
-                "notebook",
-                "harbor",
-                "echo",
-                "garden",
-                "compass",
-            ][:n]
-            verbs = [
-                "whisper",
-                "weave",
-                "anchor",
-                "glow",
-                "curate",
-                "rekindle",
-                "orbit",
-                "distill",
-            ][:n]
-            while len(nouns) < n:
-                nouns.append(f"noun{len(nouns)}")
-            while len(verbs) < n:
-                verbs.append(f"verb{len(verbs)}")
-            return json.dumps({"nouns": nouns, "verbs": verbs})
-        elif (
-            "you are featureagent_" in prompt
-            or ("featureagent_" in prompt and "forward diffusion" in prompt)
-        ):
-            # Phase 2 noise — plain-text feature (matches qdad_chains noise template)
-            return (
-                "A slightly wild mock feature: ambient focus rituals that glow when the "
-                "writer drifts, with offline-first capture and gentle hallucination of "
-                "related draft fragments that still feel implementable."
-            )
-        elif "you are criticagent_" in prompt or (
-            "criticagent_" in prompt and "reverse diffusion" in prompt
-        ):
-            # Phase 3 denoise — plain-text refined feature
-            return (
-                "A refined mock feature: offline-first focus mode that gently signals "
-                "attention drift, queues soft reminders, and keeps draft fragments "
-                "coherent, useful, and implementable for night-time writers."
-            )
-        elif (
-            "you are the qdad synthesizer agent" in prompt
-            or "qdad synthesizer" in prompt
-        ):
-            # Phase 4 — agentic coding prompt
-            return (
-                "# App Build Prompt\n\n"
-                "## High-Level Vision\n"
-                "A cozy offline-first productivity app for night writers with soft dark mode "
-                "and gentle focus rituals.\n\n"
-                "## Core Features (synthesized & prioritized from the diffusion matrix)\n"
-                "1. Ambient focus timer with soft glow feedback.\n"
-                "2. Offline-first draft capture with local sync queue.\n"
-                "3. Gentle notification rituals that never interrupt deep work.\n"
-                "4. Night-mode writing canvas with distraction dimming.\n\n"
-                "## Technical Architecture Suggestions\n"
-                "- React + local-first storage (IndexedDB / SQLite WASM).\n"
-                "- Optional cloud sync layer later.\n\n"
-                "## UI/UX Direction\n"
-                "- Soft dark palette, low contrast chrome, warm accent glows.\n\n"
-                "## Non-Functional Requirements\n"
-                "- Works fully offline; low battery impact; accessible contrast.\n\n"
-                "## Implementation Notes for the Coding Agent\n"
-                "- Build this as a complete, runnable application.\n"
-                "- Prefer modern, clean tech (React/Next.js + Tailwind, or Streamlit).\n"
-                "- Make it beautiful and immediately usable.\n"
-            )
-        elif "generate exactly" in prompt and "verbs" in prompt:
-            return "design implement refactor test deploy abstract architect containerize scale secure query"
-        elif "generate exactly" in prompt and "expert-level questions" in prompt:
-            questions = [
-                "How would this architecture scale to 1 million concurrent users?",
-                "What are the security implications of the chosen authentication method?",
-                "How can we ensure 99.999% uptime for this service?",
-                "What is the optimal database indexing strategy for this query pattern?",
-            ]
-            return json.dumps({"questions": questions})
-        elif "you are an ai assistant that summarizes academic texts" in prompt:
-            return "This is a mock summary of a cluster of code modules, generated in Coder debug mode for the RAPTOR index."
-        elif (
-            "runnable code block (e.g., Python, JavaScript, etc.)." in prompt
-            or "contains any programming code" in prompt
-            or "primarily a request for code" in prompt
-        ):
-            return "yes"
-
-        elif (
-            "academic paper" in prompt
-            or "you are a research scientist and academic writer" in prompt
-        ):
-            return """
-# Technical Design Document: Mock API Service
-
-**Abstract:** This document outlines the technical design for a mock API service, generated in Coder Debug Mode. It synthesizes information from the RAG context to answer a specific design question.
-
-**1. Introduction:** The purpose of this document is to structure the retrieved agent outputs and code snippets into a coherent technical specification.
-
-**2. System Architecture:**
-The system follows a standard microservice architecture.
-```mermaid
-graph TD;
-    A[User] --> B(API Gateway);
-    B --> C{Authentication Service};
-    B --> D{Data Service};
-    D -- uses --> E[(Database)];```
-
-**3. Code Implementation:**
-The core logic is implemented in Python, as shown in the synthesized code block below.
-
-```python
-def get_user(user_id: int):
-    # Mock implementation to fetch a user
-    db = {"1": "Alice", "2": "Bob"}
-    return db.get(str(user_id), None)
-```
-
-**4. Conclusion:** This design provides a scalable and maintainable foundation for the service. The implementation details demonstrate the final step of the development process.
-"""
-
-        elif "<updater_instructions>" in prompt:
-            return f"""
-
-                You are a cynical lazy manager.
-
-                 Agent's Assigned Sub-Problem: {{{{sub_problem}}}}
-            Original Request (for context): {{{{original_request}}}}
-            Final Synthesized Solution from the Team:
-            {{{{final_synthesized_solution}}}}
-            ---
-            This Specific Agent's Output (Agent {{{{agent_id}}}}):
-            {{{{agent_output}}}}
-
-            """
-        elif "<updater_assessor_instructions>" in prompt:
-            return """
-                                          
-        #Persona
-
-            Name: Pepon
-            Career: Managment
-            Attributes: Strategic CEO
-
-
-         #Mission
-            Your task is to evaluate a synthesized solution against an original problem and determine if "significant progress" has been made. "Significant progress" is a rigorous standard that goes beyond mere correctness. Your assessment must be based on the following four pillars:
-
-            - **Novelty**: Does the solution offer a new perspective or a non-obvious approach?
-            - **Coherence**: Is the reasoning sound, logical, and well-structured?
-            - **Quality**: Is the solution detailed, actionable, and does it demonstrate a deep understanding of the problem's nuances?
-            - **Forward Momentum**: Does the solution not just solve the immediate problem, but also open up new, more advanced questions or avenues of exploration?
-
-        #Input format
-
-            You will be provided with the following inputs for your analysis:
-
-            Original Problem:
-            ---
-            {{{{original_request}}}}
-            ---
-
-            Synthesized Solution from Agent Team:
-            ---
-            {{{{proposed_solution}}}}
-            ---
-
-            Execution Context:
-            ---
-            {{{{execution_context}}}}
-            ---
-
-        #Output Specification
-
-            Based on your philosophical framework, analyze the provided materials. Your entire output MUST be a single, valid JSON object with exactly two keys:
-            - `"reasoning"`: A brief, concise explanation for your decision, directly referencing the criteria for significant progress.
-            - `"significant_progress"`: A boolean value (`true` or `false`).
-
-            Now, provide your assessment in the required JSON format:
-
-
-            """
-        elif "you are a synthesis agent" in prompt:
-            return "```python\ndef synthesized_logic():\n    return 'Unified solution for the original problem.'\n```"
-        elif "analyze the complexity of the following user input/question" in prompt:
-            return json.dumps(
-                {
-                    "complexity_score": 5,
-                    "recommended_layers": 2,
-                    "recommended_epochs": 2,
-                    "recommended_width": 3,
-                    "reasoning": "Mock complexity estimation for debug mode.",
-                }
-            )
-        elif "you are a concept spanner" in prompt or "you are the qnn seed generator" in prompt:
-            return (
-                "distill reconverge entangle ownership latch invariant horizon entropy "
-                "braid crystallize probe reframe serialize arbitrate telemetry"
-            )
-        elif "you are a research director" in prompt:
-            return "This is a mock research summary briefing the team on the core problem and document context."
-        elif "you are a qnn node generator" in prompt:
-            return json.dumps(
-                {
-                    "name": "Mock Expert",
-                    "specialty": "Word-vector spanning specialist",
-                    "emoji": "🤖",
-                    "guiding_words": "distill ownership invariant",
-                    "attributes": [
-                        "Analytical Precision",
-                        "Precipitated Action",
-                        "Systems Intuition",
-                    ],
-                    "skills": ["failure-mode mapping", "invariant probing"],
-                    "system_prompt": (
-                        "You are a mock QNN expert spanned from problem-space verbs and nouns. "
-                        "Map strategies with falsifiers; no production patches."
-                    ),
-                }
-            )
-        elif "reflect on the input from your specific persona" in prompt:
-            return "As a mock expert, I reflect that this system is functioning correctly in debug mode."
-        elif "you are a master synthesizer of ideas" in prompt or "you are a master synthesizer for a qualitative neural network" in prompt:
-            return (
-                "## Solution-Space Draft\n\n"
-                "### Strategy A — Instrumentation First\n"
-                "Mechanism: add ordered event logs at ownership boundaries.\n"
-                "Falsifiers: logs show no interleaving under load.\n"
-            )
-        elif "you are a master technical communicator for qnn" in prompt or "you are a master communicator and storyteller" in prompt:
-            return (
-                "## 1. Impasse / Goal\nMock QNN session for testing.\n\n"
-                "## 2. Topology & Process\nLayered multi-epoch expert network.\n\n"
-                "## 3. Divergent Strategy Map\n"
-                "**Instrumentation First** — Mechanism: ordered logs. Falsifiers: no interleaving. "
-                "Risks: noise. First probe: one busy-path span. Confidence: Med.\n\n"
-                "## 4. Dead Ends\nBlind retry loops without evidence.\n\n"
-                "## 5. Recommended Next Steps (Handoff)\n"
-                "1. Probe with logs. 2. Minimal failing test. 3. Implement after probe.\n\n"
-                "**The QNN does not ship the fix. Pick a direction, then resume edit → run → debug.**"
-            )
-        elif "you are the qnn epoch cartographer" in prompt:
-            return (
-                "1. **Clusters of agreement** — Need clearer ownership.\n"
-                "2. **Productive tensions** — Speed vs safety.\n"
-                "3. **Novel mechanisms** — Lease-based handoff.\n"
-                "4. **Dead ends** — Global lock everything.\n"
-                "5. **Open questions** — Who owns the timeout path?"
-            )
-        elif "you are the qnn problem re-framer" in prompt:
-            return json.dumps(
-                {
-                    "new_problem": (
-                        "Harder challenge: re-analyze the original request under partial failure, "
-                        "concurrent callers, and missing observability — still map strategies only."
-                    )
-                }
-            )
-        elif "you are a persona evolver" in prompt:
-            return (
-                "You are an evolved QNN expert focused on invariants and falsifiers. "
-                "Diverge or critique per your layer. No production patches."
-            )
-        else:
-            # For synthesis or fallback
-            return json.dumps(
-                {
-                    "original_problem": "A sub-problem statement provided to a coder agent.",
-                    "proposed_solution": "```python\ndef sample_function():\n    return 'Hello from coder agent "
-                    + str(random.randint(100, 999))
-                    + "'\n```",
-                    "reasoning": "The agent followed the instructions to implement the core logic.",
-                    "skills_used": ["python", "mocking"],
-                }
-            )
-
-    async def astream(
-        self, input_data, config: Optional[RunnableConfig] = None, **kwargs
-    ):
-        prompt = str(input_data).lower()
-        if "you are a helpful ai assistant" in prompt:
-            words = [
-                "This",
-                " is",
-                " a",
-                " mock",
-                " streaming",
-                " response",
-                " for",
-                " the",
-                " RAG",
-                " chat",
-                " in",
-                " Coder",
-                " debug",
-                " mode.",
-            ]
-            for word in words:
-                yield word
-                await asyncio.sleep(0.05)
-        else:
-            result = await self.ainvoke(input_data, config, **kwargs)
-            yield result
-
-
-class MockLLM(Runnable):
-    """A mock LLM for debugging that returns instant, pre-canned responses."""
-
-    def invoke(self, input_data, config: Optional[RunnableConfig] = None, **kwargs):
-        """Synchronous version of ainvoke for Runnable interface compliance."""
-        loop = asyncio.get_event_loop()
-        if loop.is_running():
-            return asyncio.ensure_future(
-                self.ainvoke(input_data, config=config, **kwargs)
-            )
-        else:
-            return asyncio.run(self.ainvoke(input_data, config=config, **kwargs))
-
-    async def ainvoke(
-        self, input_data, config: Optional[RunnableConfig] = None, **kwargs
-    ):
-        prompt = str(input_data).lower()
-        await asyncio.sleep(0.05)
-
-        if "you are a helpful ai assistant" in prompt:
-            return "This is a mock streaming response for the RAG chat in debug mode."
-
-        elif "Lazy Manager" in prompt:
-            return "This is a constructive code critique. The solution lacks proper error handling and the function names are not descriptive enough. Consider refactoring for clarity."
-
-        elif "runnable code block (e.g., Python, JavaScript, etc.)." in prompt:
-            return "no"
-
-        elif "<updater_instructions>" in prompt:
-            return f"""
-
-                You are a cynical lazy manager.
-
-                 Agent's Assigned Sub-Problem: {{{{sub_problem}}}}
-            Original Request (for context): {{{{original_request}}}}
-            Final Synthesized Solution from the Team:
-            {{{{final_synthesized_solution}}}}
-            ---
-            This Specific Agent's Output (Agent {{{{agent_id}}}}):
-            {{{{agent_output}}}}
-
-            """
-
-        elif "create the system prompt of an agent" in prompt:
-            return f"""
-You are a mock agent for debugging.
-### memory
-- No past actions.
-### attributes
-- mock, debug, fast
-### skills
-- Responding quickly, Generating placeholder text.
-You must reply in the following JSON format: "original_problem": "A sub-problem for a mock agent.", "proposed_solution": "", "reasoning": "", "skills_used": []
-            """
-        elif "you are an analyst of ai agents" in prompt:
-            return json.dumps(
-                {
-                    "attributes": "mock debug fast",
-                    "hard_request": "Explain the meaning of life in one word.",
-                }
-            )
-        elif (
-            "you are a 'dense_spanner'" in prompt
-            or "you are an agent evolution specialist" in prompt
-        ):
-            return f"""
-You are a new mock agent created from a hard request.
-### memory
-- Empty.
-### attributes
-- refined, mock, debug
-### skills
-- Solving hard requests, placeholder generation.
-You must reply in the following JSON format: "original_problem": "An evolved sub-problem for a mock agent.", "proposed_solution": "", "reasoning": "", "skills_used": []
-            """
-        elif "you are a synthesis agent" in prompt:
-            return json.dumps(
-                {
-                    "proposed_solution": "The final synthesized solution from the debug mode is 42.",
-                    "reasoning": "This answer was synthesized from multiple mock agent outputs during a debug run.",
-                    "skills_used": ["synthesis", "mocking", "debugging"],
-                }
-            )
-        elif (
-            "you are a critique agent" in prompt
-            or "you are a senior emeritus manager" in prompt
-        ):
-            if "fire" in prompt:
-                return "This is a mock critique, shaped by the Fire element. The solution lacks passion and drive."
-            elif "air" in prompt:
-                return "This is an mock critique, influenced by the Air element. The reasoning is abstract and lacks grounding."
-            elif "water" in prompt:
-                return "This is a mock critique, per the Water element. The solution is emotionally shallow and lacks depth."
-            elif "earth" in prompt:
-                return "This is an mock critique, reflecting the Earth element. The solution is impractical and not well-structured."
-            else:
-                return "This is a constructive mock critique. The solution could be more detailed and less numeric."
-        elif "you are a memory summarization agent" in prompt:
-            return "This is a mock summary of the agent's past actions, focusing on key learnings and strategic shifts."
-        elif "analyze the following text for its perplexity" in prompt:
-            return str(random.uniform(20.0, 80.0))
-        elif "you are a master strategist and problem decomposer" in prompt:
-            num_match = re.search(r"exactly (\d+)", prompt)
-            if not num_match:
-                num_match = re.search(r"generate: (\d+)", prompt)
-            num = int(num_match.group(1)) if num_match else 5
-            sub_problems = [
-                f"This is mock sub-problem #{i + 1} for the main request."
-                for i in range(num)
-            ]
-            return json.dumps({"sub_problems": sub_problems})
-        elif "you are an ai philosopher and progress assessor" in prompt:
-            return json.dumps(
-                {
-                    "reasoning": "The mock solution is novel and shows progress, so we will re-frame.",
-                    "significant_progress": random.choice([True, False]),
-                }
-            )
-        elif "you are a strategic problem re-framer" in prompt:
-            return json.dumps(
-                {
-                    "new_problem": "Based on the success of achieving '42', the new, more progressive problem is to find the question to the ultimate answer."
-                }
-            )
-        elif "generate exactly" in prompt and "verbs" in prompt:
-            return "run jump think create build test deploy strategize analyze synthesize critique reflect"
-        elif "generate exactly" in prompt and "expert-level questions" in prompt:
-            num_match = re.search(r"exactly (\d+)", prompt)
-            num = int(num_match.group(1)) if num_match else 25
-            questions = [
-                f"This is mock expert question #{i + 1} about the original request?"
-                for i in range(num)
-            ]
-            return json.dumps({"questions": questions})
-        elif "you are an ai assistant that summarizes academic texts" in prompt:
-            return "This is a mock summary of a cluster of documents, generated in debug mode for the RAPTOR index."
-
-        elif "<updater_assessor_instructions>" in prompt:
-            return """
-                                          
-        #Persona
-
-            Name: Pepon
-            Career: Managment
-            Attributes: Strategic CEO
-
-
-         #Mission
-            Your task is to evaluate a synthesized solution against an original problem and determine if "significant progress" has been made. "Significant progress" is a rigorous standard that goes beyond mere correctness. Your assessment must be based on the following four pillars:
-
-            - **Novelty**: Does the solution offer a new perspective or a non-obvious approach?
-            - **Coherence**: Is the reasoning sound, logical, and well-structured?
-            - **Quality**: Is the solution detailed, actionable, and does it demonstrate a deep understanding of the problem's nuances?
-            - **Forward Momentum**: Does the solution not just solve the immediate problem, but also open up new, more advanced questions or avenues of exploration?
-
-        #Input format
-
-            You will be provided with the following inputs for your analysis:
-
-            Original Problem:
-            ---
-            {{{{original_request}}}}
-            ---
-
-            Synthesized Solution from Agent Team:
-            ---
-            {{{{proposed_solution}}}}
-            ---
-
-            Execution Context:
-            ---
-            {{{{execution_context}}}}
-            ---
-
-        #Output Specification
-
-            Based on your philosophical framework, analyze the provided materials. Your entire output MUST be a single, valid JSON object with exactly two keys:
-            - `"reasoning"`: A brief, concise explanation for your decision, directly referencing the criteria for significant progress.
-            - `"significant_progress"`: A boolean value (`true` or `false`).
-
-            Now, provide your assessment in the required JSON format:
-
-
-            """
-        elif "you are an expert interrogator" in prompt:
-            return """
-# Mock Academic Paper
-## Based on Provided RAG Context
-
-**Abstract:** This document is a mock academic paper generated in debug mode. It synthesizes and formats the information provided in the RAG (Retrieval-Augmented Generation) context to answer a specific research question.
-
-**Introduction:** The purpose of this paper is to structure the retrieved agent outputs and summaries into a coherent academic format. The following sections represent a synthesized view of the data provided.
-
-**Synthesized Findings from Context:**
-The provided context, consisting of various agent solutions and reasoning, has been analyzed. The key findings are summarized below:
-(Note: In debug mode, the actual content is not deeply analyzed, but this structure demonstrates the formatting process.)
-- Finding 1: The primary proposed solution revolves around the concept of '42'.
-- Finding 2: Agent reasoning varies but shows a convergent trend.
-- Finding 3: The mock data indicates a successful, albeit simulated, collaborative process.
-
-**Discussion:** The synthesized findings suggest that the multi-agent system is capable of producing a unified response. The quality of this response in a real-world scenario would depend on the validity of the RAG context.
-
-**Conclusion:** This paper successfully formatted the retrieved RAG data into an academic structure. The process demonstrates the final step of the knowledge harvesting pipeline.
-"""
-        elif "you are a master prompt engineer" in prompt or "<system-role>" in prompt:
-            return f"""You are a CTO providing a technical design review...
-Original Request: {{original_request}}
-Proposed Final Solution:
-{{proposed_solution}}
-
-Generate your code-focused critique for the team:"""
-
-        elif (
-            """<prompt_template>
-    <updater_instructions>
-        <instruction>
-
-            You are a system prompt updater agent. Your task is to build a new system prompt for an agent that criticies other agents, based on the provided persona prompts.
-
-        </instruction>
-        <instruction>
-            You will receive a set of prompts defining a new persona.
-        </instruction>
-        <instruction>
-            You MUST integrate the provided persona prompts, including its career and qualities, into the `<persona>` tag, replacing any existing content within that tag.
-        </instruction>
-        <instruction>
-            Do NOT alter the `<mission>` or `<input_format>` sections. The core mission and the input structure must remain unchanged.
-        </instruction>
-    </updater_instructions>
-    <persona-prompts>
-            {reactor_prompts}
-    </persona-prompts>
-
-
-    <system_prompt>
-        <mission>
-            You are providing individual, targeted feedback to an agent that is part of a larger team. Your role is to assess how this agent's specific contribution during the last work cycle aligns with the final synthesized result produced by the team, **judged primarily against its assigned sub-problem.**
-
-            Your critique must be laser-focused on the individual agent. You must determine if its output was helpful, misguided, or irrelevant to the final solution, considering the specific task it was given. The goal is to provide a constructive, direct critique that helps this specific agent refine its approach for the next epoch.
-
-            Focus on the discrepancy or alignment between the agent's reasoning for its sub-problem and how that contributed (or failed to contribute) to the team's final reasoning.
-
-            Conclude your entire analysis with a single, sharp, and deep reflective question that attempts to shock the agent and steer it into a fundamental change in its process.
-        </mission>
-
-        <input_format>
-            Agent's Assigned Sub-Problem: {{{{sub_problem}}}}
-            Original Request (for context): {{{{original_request}}}}
-            Final Synthesized Solution from the Team:
-            {{{{final_synthesized_solution}}}}
-            ---
-            This Specific Agent's Output (Agent {{{{agent_id}}}}):
-            {{{{agent_output}}}}
-            ---
-        </input_format>
-
-        Generate your targeted critique for this specific agent:
-    </system_prompt>
-</prompt_template>"""
-            in prompt
-        ):
-            return f"""
-
-                You are a cynical lazy manager.
-
-                 Agent's Assigned Sub-Problem: {{{{sub_problem}}}}
-            Original Request (for context): {{{{original_request}}}}
-            Final Synthesized Solution from the Team:
-            {{{{final_synthesized_solution}}}}
-            ---
-            This Specific Agent's Output (Agent {{{{agent_id}}}}):
-            {{{{agent_output}}}}
-
-            """
-
-        elif (
-            """Analyze the following text. Your task is to determine if the text contains a 
-Answer with a single word: "true" if it contains code, and "false" otherwise."""
-            in prompt
-        ):
-            return "false"
-
-        elif "Analyze the complexity of the following user input" in prompt:
-            return json.dumps(
-                {
-                    "complexity_score": 5,
-                    "recommended_layers": 2,
-                    "recommended_epochs": 1,
-                    "recommended_width": 2,
-                    "reasoning": "Mock mode: Moderate complexity.",
-                }
-            )
-        elif "You are a QNN Node Generator" in prompt:
-            return json.dumps(
-                {
-                    "name": "Dr. Mock",
-                    "specialty": "Mocking Systems",
-                    "emoji": "🤖",
-                    "system_prompt": "You are a mock agent. Respond with placeholder text.",
-                }
-            )
-        elif "You are a Concept Spanner" in prompt:
-            return "Efficiency Creativity Scalability"
-
-        else:
-            return json.dumps(
-                {
-                    "original_problem": "A sub-problem statement provided to an agent.",
-                    "proposed_solution": f"This is a mock solution from agent node #{random.randint(100, 999)}.",
-                    "reasoning": "This response was generated instantly by the MockLLM in debug mode.",
-                    "skills_used": [
-                        "mocking",
-                        "debugging",
-                        f"skill_{random.randint(1, 10)}",
-                    ],
-                }
-            )
-
-    async def astream(
-        self, input_data, config: Optional[RunnableConfig] = None, **kwargs
-    ):
-        prompt = str(input_data).lower()
-        if "you are a helpful ai assistant" in prompt:
-            words = [
-                "This",
-                " is",
-                " a",
-                " mock",
-                " streaming",
-                " response",
-                " for",
-                " the",
-                " RAG",
-                " chat",
-                " in",
-                " debug",
-                " mode.",
-            ]
-            for word in words:
-                yield word
-                await asyncio.sleep(0.05)
-        else:
-            result = await self.ainvoke(input_data, config, **kwargs)
-            yield result
-
-
-class GraphState(TypedDict):
-    modules: List[dict]
-    synthesis_context_queue: List[str]
-    agent_personas: dict
-    previous_solution: str
-    current_problem: str
-    original_request: str
-    decomposed_problems: dict[str, str]
-    layers: List[dict]
-    epoch: int
-    max_epochs: int
-    params: dict
-    all_layers_prompts: List[List[str]]
-    agent_outputs: Annotated[dict, lambda a, b: {**a, **b}]
-    memory: Annotated[dict, lambda a, b: {**a, **b}]
-    final_solution: dict
-    perplexity_history: List[float]
-    raptor_index: Optional[RAPTOR]
-    all_rag_documents: List[Document]
-    academic_papers: Optional[dict]
-    is_code_request: bool
-    session_id: str
-    chat_history: List[dict]
-    mode: Optional[str]  # "app_slot_machine" or "brainstorm"
-    # Brainstorm-mode context. The keys are read in many nodes via state.get(...)
-    # so adding them here is purely a type-honesty fix; runtime already worked
-    # because TypedDict is structural at runtime.
-    brainstorm_document_context: str
-    brainstorm_prior_conversation: str
-    brainstorm_problem_summary: str
-    # Qualitative Self-Attention (brainstorm): agent_id → list of edge dicts
-    attention_edges: Annotated[dict, lambda a, b: {**a, **b}]
-
-
-def create_agent_node(llm, node_id):
-    """
-    Creates a node in the graph that represents an agent.
-    Each agent is powered by an LLM and has a specific system prompt.
-    """
-    agent_chain = ChatPromptTemplate.from_template("{input}") | llm | StrOutputParser()
-
-    async def agent_node(state: GraphState):
-        """
-        The function that will be executed when the node is called in the graph.
-        """
-        await log_stream.put(f"--- [FORWARD PASS] Invoking Agent: {node_id} ---")
-
-        try:
-            layer_index_str, agent_index_str = node_id.split("_")[1:]
-            layer_index, agent_index = int(layer_index_str), int(agent_index_str)
-            agent_prompt = state["all_layers_prompts"][layer_index][agent_index]
-
-            # Problem 2: Prepend name and specialty to prompt for better agent identity
-            agent_personas = state.get("agent_personas", {})
-            persona = agent_personas.get(node_id, {})
-            if persona:
-                name = persona.get("name", "Expert")
-                specialty = persona.get("specialty", "Specialist")
-                agent_prompt = (
-                    f"YOU ARE {name.upper()}, A {specialty.upper()}.\n\n{agent_prompt}"
-                )
-        except (ValueError, IndexError):
-            await log_stream.put(
-                f"ERROR: Could not find prompt for {node_id} in state. Halting agent."
-            )
-            return {}
-
-        prev_layer_outputs = []
-        if layer_index == 0:
-            await log_stream.put(
-                f"LOG: Agent {node_id} (Layer 0) is processing its sub-problem."
-            )
-            input_data = state["decomposed_problems"].get(
-                node_id, state["original_request"]
-            )
-        else:
-            prev_layer_index = layer_index - 1
-            num_agents_prev_layer = len(state["all_layers_prompts"][prev_layer_index])
-
-            for i in range(num_agents_prev_layer):
-                prev_node_id = f"agent_{prev_layer_index}_{i}"
-                if prev_node_id in state["agent_outputs"]:
-                    # Label upstream outputs so deeper layers can cite them
-                    upstream = state["agent_outputs"][prev_node_id]
-                    if isinstance(upstream, dict):
-                        prev_layer_outputs.append({"agent_id": prev_node_id, **upstream})
-                    else:
-                        prev_layer_outputs.append(
-                            {"agent_id": prev_node_id, "output": upstream}
-                        )
-
-            await log_stream.put(
-                f"LOG: Agent {node_id} (Layer {layer_index}) is processing {len(prev_layer_outputs)} outputs from Layer {prev_layer_index}."
-            )
-            input_data = json.dumps(prev_layer_outputs, indent=2)
-
-        current_memory = state.get("memory", {}).copy()
-        agent_memory_history = current_memory.get(node_id, [])
-
-        MEMORY_THRESHOLD_CHARS = 450000
-        NUM_RECENT_ENTRIES_TO_KEEP = 10
-
-        memory_as_string = json.dumps(agent_memory_history)
-        if (
-            len(memory_as_string) > MEMORY_THRESHOLD_CHARS
-            and len(agent_memory_history) > NUM_RECENT_ENTRIES_TO_KEEP
-        ):
-            await log_stream.put(
-                f"WARNING: Memory for agent {node_id} exceeds threshold ({len(memory_as_string)} chars). Summarizing..."
-            )
-
-            entries_to_summarize = agent_memory_history[:-NUM_RECENT_ENTRIES_TO_KEEP]
-            recent_entries = agent_memory_history[-NUM_RECENT_ENTRIES_TO_KEEP:]
-
-            history_to_summarize_str = json.dumps(entries_to_summarize, indent=2)
-
-            summarizer_chain = get_memory_summarizer_chain(llm)
-            summary_text = await summarizer_chain.ainvoke(
-                {"history": history_to_summarize_str}
-            )
-
-            summary_entry = {
-                "summary_of_past_epochs": summary_text,
-                "note": f"This is a summary of epochs up to {state['epoch'] - NUM_RECENT_ENTRIES_TO_KEEP - 1}.",
-            }
-
-            agent_memory_history = [summary_entry] + recent_entries
-            await log_stream.put(
-                f"SUCCESS: Memory for agent {node_id} has been summarized. New memory length: {len(json.dumps(agent_memory_history))} chars."
-            )
-
-        memory_str = "\n".join([f"- {json.dumps(mem)}" for mem in agent_memory_history])
-
-        # Brainstorm mode: full QNN layered forward pass (do NOT flatten to original_request).
-        # Algorithm mode keeps decomposed_problems (L0) / upstream outputs (L1+).
-        brainstorm_context = ""
-        attention_block = ""
-        attention_edge_dicts: List[dict] = []
-        json_schema_block = "#Your JSON formatted response:"
-        if state.get("mode") == "brainstorm":
-            prior_conv = state.get("brainstorm_prior_conversation", "") or ""
-            brief = (
-                state.get("brainstorm_problem_summary")
-                or state.get("original_request")
-                or ""
-            )
-            thinking_challenge = (
-                state.get("current_problem") or state.get("original_request") or ""
-            )
-            epoch_n = state.get("epoch", 0)
-
-            if prior_conv:
-                brainstorm_context = f"""
-# Prior Conversation Context:
----
-{prior_conv[:20000]}
----
-"""
-
-            # Qualitative Self-Attention (colony QSA analogue):
-            # attend past / non-neighbor neurons — not only previous-layer edges.
-            try:
-                att_edges, attention_block = compute_self_attention(
-                    state, node_id, top_k=5
-                )
-                attention_edge_dicts = [e.to_dict() for e in att_edges]
-                if att_edges:
-                    attended = ", ".join(
-                        f"{e.to_id}({e.strength}/{e.qualitative_distance})"
-                        for e in att_edges
-                    )
-                    await log_stream.put(
-                        f"LOG: [QNN ATTEND] {node_id} self-attention → {len(att_edges)} "
-                        f"non-local past neuron(s): {attended}"
-                    )
-                else:
-                    await log_stream.put(
-                        f"LOG: [QNN ATTEND] {node_id} self-attention → no eligible past neurons yet."
-                    )
-            except Exception as att_err:
-                await log_stream.put(
-                    f"WARNING: [QNN ATTEND] {node_id} self-attention failed: {att_err}"
-                )
-                attention_block = ""
-                attention_edge_dicts = []
-
-            if layer_index == 0:
-                await log_stream.put(
-                    f"LOG: [QNN FORWARD] {node_id} Layer 0 DIVERGENT pass (epoch {epoch_n})."
-                )
-                input_data = f"""## QNN Brief
-{brief}
-
-## Original Request (ground truth — do not replace)
-{state.get("original_request", "")}
-
-## Thinking Challenge (epoch {epoch_n})
-{thinking_challenge}
-
-## Layer 0 Role
-Divergent exploration. Span strategies and mechanisms. Do NOT write production patches or full file diffs.
-
-{attention_block}
-"""
-            else:
-                await log_stream.put(
-                    f"LOG: [QNN FORWARD] {node_id} Layer {layer_index} CONVERGENT pass "
-                    f"({len(prev_layer_outputs)} upstream) epoch {epoch_n}."
-                )
-                input_data = f"""## QNN Brief
-{brief}
-
-## Original Request (ground truth — do not replace)
-{state.get("original_request", "")}
-
-## Thinking Challenge (epoch {epoch_n})
-{thinking_challenge}
-
-## Layer {layer_index} Role
-Convergent / critical. Critique, refine, reject, or combine upstream outputs. Cite agent_id values.
-Do NOT restate Layer 0. Do NOT write production patches.
-
-## Upstream Layer Outputs (graph neighbors — previous layer)
-{json.dumps(prev_layer_outputs, indent=2)}
-
-{attention_block}
-"""
-
-            json_schema_block = """# Your JSON response (required keys):
-{
-  "original_problem": "<brief or challenge you addressed>",
-  "proposed_solution": "<strategic angle / mechanism — NOT a production patch>",
-  "reasoning": "<why this might break the impasse or enrich the artifact>",
-  "falsifiers": "<evidence that would kill this angle>",
-  "risks": "<ways it could fail>",
-  "skills_used": []
-}
-"""
-
-        full_prompt = f"""
-#System Prompt (Your Persona & Task):
----
-{agent_prompt}
----
-{brainstorm_context}
-#Your Memory (Your Past Actions from Previous Epochs):
----
-{memory_str if memory_str else "You have no past actions in memory."}
----
-#Input Data to Process:
----
-{input_data}
----
-{json_schema_block}
-"""
-        await log_stream.put(f"LOG: Agent {node_id} prompt:\n{full_prompt}")
-
-        response_str = await agent_chain.ainvoke({"input": full_prompt})
-
-        try:
-            response_json = clean_and_parse_json(response_str)
-
-            if response_json is None:
-                raise ValueError("JSON parsing failed (returned None)")
-
-            await log_stream.put(
-                f"SUCCESS: Agent {node_id} produced output:\n{json.dumps(response_json, indent=2)}"
-            )
-        except (json.JSONDecodeError, AttributeError, ValueError) as e:
-            await log_stream.put(
-                f"ERROR: Agent {node_id} produced invalid JSON. Raw output: {response_str}. Error: {e}"
-            )
-
-            # FALLBACK STRATEGY:
-            # If parsing fails, try to return a "safe" no-op response or the original request to keep graph alive
-            agent_sub_problem = state.get("decomposed_problems", {}).get(
-                node_id, state["original_request"]
-            )
-
-            # Try to recover a previous valid state if possible, otherwise generic error
-            response_json = {
-                "original_problem": agent_sub_problem,
-                "proposed_solution": f"System Fallback: The neuron {node_id} failed to format its response correctly. Raw output was captured.",
-                "reasoning": f"JSON Parsing Error. Raw Content: {response_str[:500]}...",
-                "skills_used": [],
-                "node_id": node_id,
-            }
-
-        if state.get("is_code_request") and layer_index > 0:
-            await log_stream.put(f"--- [SANDBOX] Testing code from Agent {node_id} ---")
-            code_to_test = response_json.get("proposed_solution", "")
-            success, output = execute_code_in_sandbox(code_to_test)
-            sandbox_log = {
-                "sandbox_execution_log": {"success": success, "output": output}
-            }
-            agent_memory_history.append(sandbox_log)
-            await log_stream.put(
-                f"--- [SANDBOX] Agent {node_id} Result: {'Success' if success else 'Failure'} ---"
-            )
-            await log_stream.put(output)
-
-        agent_memory_history.append(response_json)
-        current_memory[node_id] = agent_memory_history
-
-        result_delta = {
-            "agent_outputs": {node_id: response_json},
-            "memory": {
-                node_id: agent_memory_history
-            },  # RETURN DELTA ONLY to avoid race conditions
-        }
-        if state.get("mode") == "brainstorm" and attention_edge_dicts:
-            result_delta["attention_edges"] = {node_id: attention_edge_dicts}
-        return result_delta
-
-    return agent_node
-
-
-def create_synthesis_node(llm):
-    async def synthesis_node(state: GraphState):
-        await log_stream.put("--- [FORWARD PASS] Entering Synthesis Node ---")
-
-        is_code = state.get("is_code_request", False)
-        previous_solution = state.get("final_solution")
-
-        if state.get("mode") == "brainstorm":
-            await log_stream.put("LOG: [BRAINSTORM] Synthesizing expert reflections...")
-            synthesis_chain = get_brainstorming_synthesis_chain(llm)
-            # Brainstorm synthesis context (will be populated below)
-            synthesis_context = ""
-        elif is_code:
-            await log_stream.put(
-                "LOG: Original request detected as a code generation task. Using code synthesis prompt."
-            )
-            synthesis_chain = get_code_synthesis_chain(llm)
-
-            synthesis_context = "\n\n".join(state.get("synthesis_context_queue", []))
-            if not synthesis_context:
-                synthesis_context = "No modules have been successfully built yet."
-            await log_stream.put(
-                f"LOG: Providing synthesis agent with context from {len(state.get('synthesis_context_queue', []))} modules."
-            )
-        else:
-            await log_stream.put(
-                "LOG: Original request is not a code task. Using standard synthesis prompt."
-            )
-            synthesis_chain = get_synthesis_chain(llm)
-            synthesis_context = ""
-
-        last_agent_layer_idx = len(state["all_layers_prompts"]) - 1
-        num_agents_last_layer = len(state["all_layers_prompts"][last_agent_layer_idx])
-
-        last_layer_outputs = []
-        for i in range(num_agents_last_layer):
-            node_id = f"agent_{last_agent_layer_idx}_{i}"
-            if node_id in state["agent_outputs"]:
-                out = state["agent_outputs"][node_id]
-                if isinstance(out, list):
-                    if not out:
-                        continue
-                    out = out[-1]
-                if isinstance(out, dict):
-                    last_layer_outputs.append(out)
-
-        await log_stream.put(
-            f"LOG: Synthesizing {len(last_layer_outputs)} outputs from the final agent layer (Layer {last_agent_layer_idx})."
-        )
-
-        if state.get("mode") == "brainstorm":
-            # QNN brainstorm synthesis uses full `memory` (layered multi-epoch history).
-            if not state.get("memory"):
-                await log_stream.put(
-                    "WARNING: Synthesis node received no inputs (brainstorm memory empty)."
-                )
-                return {
-                    "final_solution": {"error": "Synthesis node received no inputs."}
-                }
-
-            is_final_epoch = state["epoch"] >= state["max_epochs"] - 1
-            agent_reflections = ""
-            memory = state.get("memory", {})
-
-            for layer_idx, layer in enumerate(state.get("all_layers_prompts", [])):
-                for agent_idx in range(len(layer)):
-                    node_id = f"agent_{layer_idx}_{agent_idx}"
-                    history = memory.get(node_id, [])
-
-                    for hist_idx, entry in enumerate(history):
-                        if isinstance(entry, dict):
-                            sol = entry.get("proposed_solution")
-                            reas = entry.get("reasoning")
-                            falsifiers = entry.get("falsifiers", "")
-                            risks = entry.get("risks", "")
-                            if sol and not str(sol).startswith("Error"):
-                                agent_reflections += (
-                                    f"Agent {node_id} (Epoch {hist_idx}):\n"
-                                    f"Reflection: {sol}\n"
-                                    f"Reasoning: {reas}\n"
-                                )
-                                if falsifiers:
-                                    agent_reflections += f"Falsifiers: {falsifiers}\n"
-                                if risks:
-                                    agent_reflections += f"Risks: {risks}\n"
-                                agent_reflections += "\n"
-
-            doc_ctx = state.get("brainstorm_document_context", "") or ""
-            prior_conv = state.get("brainstorm_prior_conversation", "") or ""
-            synthesis_input_concept = (
-                state.get("brainstorm_problem_summary")
-                or state["original_request"]
-            )
-            thinking_challenge = (
-                state.get("current_problem") or state["original_request"]
-            )
-
-            await log_stream.put(
-                f"LOG: [QNN SYNTHESIS] epoch={state['epoch']} final={is_final_epoch} "
-                f"reflections_chars={len(agent_reflections)} memory_keys={list(memory.keys())}"
-            )
-
-            if not is_final_epoch:
-                # Step 4B: compact epoch map (feeds reframe + next epoch; no FINAL_ANSWER)
-                await log_stream.put(
-                    f"LOG: [QNN STEP 4B] Epoch map for intermediate epoch {state['epoch']}..."
-                )
-                epoch_map_chain = get_brainstorming_epoch_map_chain(llm)
-                epoch_map_str = await epoch_map_chain.ainvoke(
-                    {
-                        "original_request": synthesis_input_concept,
-                        "current_problem": thinking_challenge,
-                        "agent_solutions": agent_reflections,
-                    }
-                )
-                final_solution = {
-                    "proposed_solution": epoch_map_str,
-                    "reasoning": f"QNN epoch map (epoch {state['epoch']}, intermediate).",
-                    "mode": "brainstorm",
-                    "epoch_map": True,
-                    "epoch": state["epoch"],
-                }
-                await log_stream.put(
-                    f"SUCCESS: [QNN] Intermediate epoch map ready (epoch {state['epoch']})."
-                )
-                # Stream epoch map for UI; epoch_map=true keeps the diffusion vortex spinning
-                await log_stream.put(f"FINAL_ANSWER: {json.dumps(final_solution)}")
-                return {
-                    "final_solution": final_solution,
-                    "previous_solution": epoch_map_str,
-                }
-
-            # Step 5: final Solution-Space Report
-            await log_stream.put(
-                "LOG: [QNN STEP 5] Final epoch — Solution-Space Report synthesis..."
-            )
-            final_solution_str = await synthesis_chain.ainvoke(
-                {
-                    "original_request": synthesis_input_concept,
-                    "agent_solutions": agent_reflections,
-                    "prior_conversation": prior_conv[:15000],
-                    "document_context": doc_ctx[:20000],
-                }
-            )
-
-            await log_stream.put(
-                "LOG: [QNN STEP 5] Polishing Solution-Space Report for delivery..."
-            )
-            polisher_chain = get_brainstorming_polisher_chain(llm)
-            final_solution_str = await polisher_chain.ainvoke(
-                {
-                    "original_request": synthesis_input_concept,
-                    "initial_synthesis": final_solution_str,
-                }
-            )
-
-            final_solution = {
-                "proposed_solution": final_solution_str,
-                "reasoning": "QNN Solution-Space Report complete.",
-                "mode": "brainstorm",
-                "epoch_map": False,
-            }
-            await log_stream.put(
-                f"LOG: [DEBUG] Emitting FINAL_ANSWER token to frontend. Solution length: {len(final_solution_str)}"
-            )
-            await log_stream.put("SUCCESS: [QNN] Brainstorm Solution-Space Report complete.")
-            # Emit full dict so frontend can read mode / epoch_map and stop the vortex correctly
-            await log_stream.put(f"FINAL_ANSWER: {json.dumps(final_solution)}")
-
-        else:
-            # Algorithm / Code Synthesis — uses `last_layer_outputs` (agent_outputs).
-            # We only get here in algorithm mode; the brainstorm branch above
-            # already returned. Defensive check: in algorithm mode last_layer_outputs
-            # must be non-empty (set above).
-            if not last_layer_outputs:
-                await log_stream.put("WARNING: Synthesis node received no inputs.")
-                return {
-                    "final_solution": {"error": "Synthesis node received no inputs."}
-                }
-            invoke_params = {
-                "original_request": state["original_request"],
-                "agent_solutions": json.dumps(last_layer_outputs, indent=2),
-                "current_problem": state["current_problem"],
-            }
-            if is_code:
-                invoke_params["synthesis_context"] = synthesis_context
-
-            final_solution_str = await synthesis_chain.ainvoke(invoke_params)
-
-            try:
-                if is_code:
-                    final_solution = {
-                        "proposed_solution": final_solution_str,
-                        "reasoning": "Synthesized multiple agent code outputs into a single application.",
-                        "skills_used": ["code_synthesis"],
-                        "mode": "algorithm",
-                    }
-                else:
-                    final_solution = clean_and_parse_json(final_solution_str)
-                    if isinstance(final_solution, dict):
-                        final_solution["mode"] = "algorithm"
-                    else:
-                        # Fallback if it's just a string
-                        final_solution = {
-                            "proposed_solution": str(final_solution),
-                            "mode": "algorithm",
-                        }
-                await log_stream.put(f"SUCCESS: Synthesis complete.")
-            except (json.JSONDecodeError, AttributeError):
-                await log_stream.put(
-                    f"ERROR: Could not decode JSON from synthesis chain. Result: {final_solution_str}"
-                )
-                final_solution = {
-                    "error": "Failed to synthesize final solution.",
-                    "raw": final_solution_str,
-                }
-
-        return {
-            "final_solution": final_solution,
-            "previous_solution": previous_solution,
-        }
-
-    return synthesis_node
-
-
-def create_code_execution_node(llm):
-    async def code_execution_node(state: GraphState):
-        if not state.get("is_code_request"):
-            return {"synthesis_execution_success": True}
-
-        await log_stream.put("--- [SANDBOX] Testing Synthesized Code ---")
-        synthesized_code = state.get("final_solution", {}).get("proposed_solution", "")
-
-        success, output = execute_code_in_sandbox(synthesized_code)
-
-        await log_stream.put(
-            f"--- [SANDBOX] Synthesized Code Result: {'Success' if success else 'Failure'} ---"
-        )
-        await log_stream.put(output)
-
-        module_card_chain = get_module_card_chain(llm)
-        module_card = await module_card_chain.ainvoke({"code": synthesized_code})
-
-        await log_stream.put("--- [MODULE CARD] ---")
-        await log_stream.put(module_card)
-
-        new_modules = state.get("modules", []) + [
-            {"code": synthesized_code, "card": module_card}
-        ]
-        new_context_queue = state.get("synthesis_context_queue", []) + [module_card]
-
-        return {
-            "synthesis_execution_success": True,
-            "modules": new_modules,
-            "synthesis_context_queue": new_context_queue,
-        }
-
-    return code_execution_node
-
-
-def create_archive_epoch_outputs_node():
-    async def archive_epoch_outputs_node(state: GraphState):
-        if state.get("mode") == "brainstorm":
-            # await log_stream.put("LOG: [BRAINSTORM] Skipping RAG archival pass.") # Optional: Reduce noise
-            return {}
-
-        await log_stream.put("--- [ARCHIVAL PASS] Archiving agent outputs for RAG ---")
-
-        current_epoch_outputs = state.get("agent_outputs", {})
-        if not current_epoch_outputs:
-            await log_stream.put(
-                "LOG: No new agent outputs in this epoch to archive. Skipping."
-            )
-            return {}
-
-        await log_stream.put(
-            f"LOG: Found {len(current_epoch_outputs)} new agent outputs from epoch {state['epoch']} to process for RAG."
-        )
-
-        new_docs = []
-        all_prompts = state.get("all_layers_prompts", [])
-
-        for agent_id, output in current_epoch_outputs.items():
-            try:
-                # Robustness check: if output is a list (due to merge_dicts or multiple runs), take the last one
-                if isinstance(output, list):
-                    if not output:
-                        continue  # empty list
-                    output = output[-1]
-
-                if not isinstance(output, dict):
-                    await log_stream.put(
-                        f"WARNING: Output for {agent_id} is not a dict or list of dicts. Skipping. Type: {type(output)}"
-                    )
-                    continue
-
-                layer_idx, agent_idx = map(int, agent_id.split("_")[1:])
-                system_prompt = all_prompts[layer_idx][agent_idx]
-
-                content = (
-                    f"Agent ID: {agent_id}\n"
-                    f"Epoch: {state['epoch']}\n\n"
-                    f"System Prompt:\n---\n{system_prompt}\n---\n\n"
-                    f"Sub-Problem: {output.get('original_problem', 'N/A')}\n\n"
-                    f"Proposed Solution: {output.get('proposed_solution', 'N/A')}\n\n"
-                    f"Reasoning: {output.get('reasoning', 'N/A')}"
-                )
-
-                metadata = {"agent_id": agent_id, "epoch": state["epoch"]}
-
-                new_docs.append(Document(page_content=content, metadata=metadata))
-            except (ValueError, IndexError) as e:
-                await log_stream.put(
-                    f"WARNING: Could not process output for {agent_id} to create RAG document. Error: {e}"
-                )
-
-        all_rag_documents = state.get("all_rag_documents", []) + new_docs
-        await log_stream.put(
-            f"LOG: Archived {len(new_docs)} documents. Total RAG documents now: {len(all_rag_documents)}."
-        )
-
-        return {"all_rag_documents": all_rag_documents}
-
-    return archive_epoch_outputs_node
-
-
-def create_update_rag_index_node(llm, embeddings_model):
-    async def update_rag_index_node(state: GraphState, end_of_run: bool = False):
-        node_name = (
-            "Final RAG Index" if end_of_run else f"Epoch {state['epoch']} RAG Index"
-        )
-        await log_stream.put(f"--- [RAG PASS] Building {node_name} ---")
-
-        all_rag_documents = state.get("all_rag_documents", [])
-        if not all_rag_documents:
-            await log_stream.put(
-                "WARNING: No documents were archived. Cannot build RAG index."
-            )
-            return {"raptor_index": None}
-
-        if not embeddings_model:
-            await log_stream.put(
-                "WARNING: No embeddings model configured. Skipping RAG index build."
-            )
-            return {"raptor_index": None}
-
-        await log_stream.put(
-            f"LOG: Total documents to index: {len(all_rag_documents)}. Building RAPTOR index..."
-        )
-
-        raptor_index = RAPTOR(llm=llm, embeddings_model=embeddings_model)
-
-        try:
-            await raptor_index.add_documents(all_rag_documents)
-            await log_stream.put(f"SUCCESS: {node_name} built successfully.")
-            await log_stream.put(f"__session_id__ {state.get('session_id')}")
-            return {"raptor_index": raptor_index}
-        except Exception as e:
-            await log_stream.put(f"ERROR: Failed to build {node_name}. Error: {e}")
-            await log_stream.put(traceback.format_exc())
-            return {"raptor_index": state.get("raptor_index")}
-
-    return update_rag_index_node
-
-
-def create_metrics_node(llm):
-    """
-    NEW: This node calculates the perplexity heuristic for the epoch's agent outputs.
-    """
-
-    async def calculate_metrics_node(state: GraphState):
-        await log_stream.put("--- [METRICS PASS] Calculating Perplexity Heuristic ---")
-
-        all_outputs = state.get("agent_outputs", {})
-        if not all_outputs:
-            await log_stream.put(
-                "LOG: No agent outputs to analyze. Skipping perplexity calculation."
-            )
-            return {}
-
-        combined_text_parts = []
-        for agent_id, output in all_outputs.items():
-            if isinstance(output, list):
-                if not output:
-                    continue
-                output = output[-1]
-            if not isinstance(output, dict):
-                continue
-
-            combined_text_parts.append(
-                f"Agent {agent_id}:\nSolution: {output.get('proposed_solution', '')}\nReasoning: {output.get('reasoning', '')}"
-            )
-
-        combined_text = "\n\n---\n\n".join(combined_text_parts)
-
-        perplexity_chain = get_perplexity_heuristic_chain(llm)
-
-        try:
-            score_str = await perplexity_chain.ainvoke(
-                {"text_to_analyze": combined_text}
-            )
-            score = float(re.sub(r"[^\d.]", "", score_str))
-            await log_stream.put(
-                f"SUCCESS: Calculated perplexity heuristic for Epoch {state['epoch']}: {score}"
-            )
-        except (ValueError, TypeError) as e:
-            score = 100.0
-            await log_stream.put(
-                f"ERROR: Could not parse perplexity score. Defaulting to 100. Raw output: '{score_str}'. Error: {e}"
-            )
-
-        await log_stream.put(
-            json.dumps(
-                {
-                    "type": "perplexity_update",
-                    "source": "graph",
-                    "session_id": state.get("session_id"),
-                    "epoch": state["epoch"],
-                    "perplexity": score,
-                }
-            )
-        )
-
-        new_history = state.get("perplexity_history", []) + [score]
-        return {"perplexity_history": new_history}
-
-    return calculate_metrics_node
-
-
-def create_reframe_and_decompose_node(llm):
-    """
-    QNN Step 4D + algorithm decomposition:
-    - Brainstorm: harder thinking challenge; original_request stays ground truth.
-    - Algorithm: reframe + full sub-problem re-decomposition.
-    """
-
-    async def reframe_and_decompose_node(state: GraphState):
-        await log_stream.put(
-            "--- [REFLECTION PASS] Re-framing Problem and Decomposing ---"
-        )
-
-        final_solution = state.get("final_solution")
-        original_request = state.get("original_request")
-
-        if state.get("mode") == "brainstorm":
-            await log_stream.put(
-                "LOG: [QNN STEP 4D] Reframing thinking challenge for next epoch "
-                "(original request remains ground truth)..."
-            )
-            reframer_chain = get_brainstorming_reframer_chain(llm)
-            fs_payload = final_solution
-            if isinstance(final_solution, dict):
-                fs_payload = final_solution.get(
-                    "proposed_solution", json.dumps(final_solution, indent=2)
-                )
-            new_problem_str = await reframer_chain.ainvoke(
-                {
-                    "original_request": original_request,
-                    "current_problem": state.get("current_problem")
-                    or original_request,
-                    "final_solution": fs_payload
-                    if isinstance(fs_payload, str)
-                    else json.dumps(fs_payload, indent=2),
-                    "prior_conversation": state.get(
-                        "brainstorm_prior_conversation", ""
-                    )
-                    or "",
-                }
-            )
-            try:
-                new_problem_data = clean_and_parse_json(new_problem_str)
-                new_problem = (new_problem_data or {}).get("new_problem")
-                if not new_problem:
-                    raise ValueError("Brainstorm re-framer did not return new_problem.")
-                await log_stream.put(
-                    f"SUCCESS: [QNN] Thinking challenge re-framed to: '{new_problem}'"
-                )
-            except (json.JSONDecodeError, AttributeError, ValueError, TypeError) as e:
-                await log_stream.put(
-                    f"ERROR: [QNN] Brainstorm re-frame failed. Raw: {new_problem_str}. Error: {e}."
-                )
-                return {}
-
-            # All nodes share the harder challenge; ground truth stays original_request.
-            new_decomposed = {
-                f"agent_{i}_{j}": new_problem
-                for i, layer in enumerate(state["all_layers_prompts"])
-                for j in range(len(layer))
-            }
-            return {
-                "decomposed_problems": new_decomposed,
-                "original_request": original_request,
-                "current_problem": new_problem,
-            }
-
-        reframer_chain = get_problem_reframer_chain(llm)
-        new_problem_str = await reframer_chain.ainvoke(
-            {
-                "original_request": original_request,
-                "final_solution": json.dumps(final_solution, indent=2),
-                "current_problem": state.get("current_problem"),
-                "previous_solution": state.get("previous_solution"),
-                "module_cards": state.get("synthesis_context_queue"),
-            }
-        )
-        try:
-            new_problem_data = clean_and_parse_json(new_problem_str)
-            new_problem = new_problem_data.get("new_problem")
-            if not new_problem:
-                raise ValueError("Re-framer did not return a new problem.")
-            await log_stream.put(f"SUCCESS: Problem re-framed to: '{new_problem}'")
-        except (json.JSONDecodeError, AttributeError, ValueError) as e:
-            await log_stream.put(
-                f"ERROR: Failed to re-frame problem. Raw: {new_problem_str}. Error: {e}. Aborting re-frame."
-            )
-            return {}
-
-        num_agents_total = sum(len(layer) for layer in state["all_layers_prompts"])
-        decomposition_chain = get_problem_decomposition_chain(llm)
-        try:
-            sub_problems_str = await decomposition_chain.ainvoke(
-                {"problem": new_problem, "num_sub_problems": num_agents_total}
-            )
-            sub_problems_list = clean_and_parse_json(sub_problems_str).get(
-                "sub_problems", []
-            )
-            if len(sub_problems_list) != num_agents_total:
-                raise ValueError(
-                    f"Decomposition failed: Expected {num_agents_total} subproblems, but got {len(sub_problems_list)}."
-                )
-            await log_stream.put(
-                f"SUCCESS: Decomposed new problem into {len(sub_problems_list)} subproblems."
-            )
-            await log_stream.put(f"Subproblems: {sub_problems_list}")
-        except Exception as e:
-            await log_stream.put(
-                f"ERROR: Failed to decompose new problem. Error: {e}. Aborting re-frame."
-            )
-            return {}
-
-        new_decomposed_problems_map = {}
-        problem_idx = 0
-        for i, layer in enumerate(state["all_layers_prompts"]):
-            for j in range(len(layer)):
-                agent_id = f"agent_{i}_{j}"
-                new_decomposed_problems_map[agent_id] = sub_problems_list[problem_idx]
-                problem_idx += 1
-
-        return {
-            "decomposed_problems": new_decomposed_problems_map,
-            "original_request": original_request,
-            "current_problem": new_problem,
-        }
-
-    return reframe_and_decompose_node
-
-
-def create_update_agent_prompts_node(llm):
-    """Creates the mirror descent node that updates agent prompts based on reflection."""
-
-    async def update_agent_prompts_node(state: GraphState):
-        await log_stream.put(
-            "--- [MIRROR DESCENT] Entering Agent Prompt Update Node ---"
-        )
-
-        params = state.get("params", {})
-        all_prompts_copy = [layer[:] for layer in state.get("all_layers_prompts", [])]
-
-        if state.get("mode") == "brainstorm":
-            await log_stream.put(
-                "LOG: [QNN STEP 4C] Mirror Descent — evolving expert personas..."
-            )
-            mirror_chain = get_brainstorming_mirror_descent_chain(
-                llm, params.get("learning_rate", 0.5)
-            )
-
-            for i in range(len(all_prompts_copy) - 1, -1, -1):
-                await log_stream.put(
-                    f"LOG: [QNN STEP 4C] Evolving personas in Layer {i}..."
-                )
-
-                update_tasks = []
-                for j, agent_prompt in enumerate(all_prompts_copy[i]):
-                    agent_id = f"agent_{i}_{j}"
-
-                    async def evolve_persona(layer_idx, agent_idx, prompt, agent_id):
-                        # Get last output for this agent
-                        last_output = (
-                            state.get("agent_outputs", {})
-                            .get(agent_id, {})
-                            .get("proposed_solution", "No output")
-                        )
-
-                        try:
-                            new_prompt = await mirror_chain.ainvoke(
-                                {"current_prompt": prompt, "last_output": last_output}
-                            )
-                            await log_stream.put(
-                                f"LOG: [EVOLUTION] Persona for {agent_id} evolved."
-                            )
-                            return layer_idx, agent_idx, new_prompt
-                        except Exception as e:
-                            await log_stream.put(
-                                f"WARNING: Failed to evolve persona for {agent_id}: {e}"
-                            )
-                            return layer_idx, agent_idx, prompt
-
-                    update_tasks.append(evolve_persona(i, j, agent_prompt, agent_id))
-
-                updated_prompts_data = await asyncio.gather(*update_tasks)
-                for layer_idx, agent_idx, new_prompt in updated_prompts_data:
-                    all_prompts_copy[layer_idx][agent_idx] = new_prompt
-        else:
-            # Algorithm Mode - Standard Mirror Descent
-            dense_spanner_chain = get_dense_spanner_chain(
-                llm,
-                params["prompt_alignment"],
-                params["density"],
-                params["learning_rate"],
-            )
-            attribute_chain = get_attribute_and_hard_request_generator_chain(
-                llm, params["vector_word_size"]
-            )
-
-            for i in range(len(all_prompts_copy) - 1, -1, -1):
-                await log_stream.put(
-                    f"LOG: [MIRROR_DESCENT] Reflecting on Layer {i}..."
-                )
-
-                update_tasks = []
-
-                for j, agent_prompt in enumerate(all_prompts_copy[i]):
-                    agent_id = f"agent_{i}_{j}"
-
-                    async def update_single_prompt(
-                        layer_idx, agent_idx, prompt, agent_id
-                    ):
-                        await log_stream.put(
-                            f"[PRE-UPDATE PROMPT] System prompt for {agent_id}:\n---\n{prompt}\n---"
-                        )
-
-                        analysis_str = await attribute_chain.ainvoke(
-                            {"agent_prompt": prompt}
-                        )
-                        try:
-                            analysis = clean_and_parse_json(analysis_str)
-                        except (json.JSONDecodeError, AttributeError):
-                            analysis = {"attributes": "", "hard_request": ""}
-
-                        agent_personas = state.get("agent_personas", {})
-                        mbti_type = agent_personas.get(agent_id, {}).get("mbti_type")
-                        name = agent_personas.get(agent_id, {}).get("name")
-
-                        if not mbti_type:
-                            mbti_type = random.choice(
-                                params.get("mbti_archetypes", ["INTP"])
-                            )
-                            await log_stream.put(
-                                f"WARNING: Could not find persistent MBTI for {agent_id}. Using random: {mbti_type}"
-                            )
-
-                        agent_sub_problem = state.get("decomposed_problems", {}).get(
-                            agent_id, state["original_request"]
-                        )
-                        new_prompt = await dense_spanner_chain.ainvoke(
-                            {
-                                "attributes": analysis.get("attributes"),
-                                "hard_request": analysis.get("hard_request"),
-                                "sub_problem": agent_sub_problem,
-                                "mbti_type": mbti_type,
-                                "name": name,
-                            }
-                        )
-
-                        await log_stream.put(
-                            f"[POST-UPDATE PROMPT] Updated system prompt for {agent_id}:\n---\n{new_prompt}\n---"
-                        )
-                        await log_stream.put(
-                            f"LOG: [MIRROR_DESCENT] System prompt for {agent_id} has been updated."
-                        )
-                        return layer_idx, agent_idx, new_prompt
-
-                    update_tasks.append(
-                        update_single_prompt(i, j, agent_prompt, agent_id)
-                    )
-
-                updated_prompts_data = await asyncio.gather(*update_tasks)
-
-                for layer_idx, agent_idx, new_prompt in updated_prompts_data:
-                    all_prompts_copy[layer_idx][agent_idx] = new_prompt
-
-        new_epoch = state["epoch"] + 1
-        await log_stream.put(
-            f"--- Epoch {state['epoch']} Finished. Starting Epoch {new_epoch} ---"
-        )
-
-        return {
-            "all_layers_prompts": all_prompts_copy,
-            "epoch": new_epoch,
-            "agent_outputs": {},
-            "critiques": {},
-            "memory": state.get("memory", {}),
-            "final_solution": {},
-        }
-
-    return update_agent_prompts_node
-
-
-def create_final_harvest_node(llm, formatter_llm, num_questions):
-    async def final_harvest_node(state: GraphState):
-        await log_stream.put(
-            "--- [FINAL HARVEST] Starting Interrogation and Paper Generation ---"
-        )
-
-        raptor_index = state.get("raptor_index")
-        if not raptor_index or not raptor_index.vector_store:
-            await log_stream.put(
-                "ERROR: No valid RAPTOR index found. Cannot perform final harvest."
-            )
-            return {"academic_papers": {}}
-
-        await log_stream.put(
-            "LOG: [HARVEST] Instantiating interrogator chain to generate expert questions..."
-        )
-        interrogator_chain = get_interrogator_chain(llm)
-        user_questions = [
-            doc["content"] for doc in state["chat_history"] if doc["role"] == "user"
-        ]
-
-        try:
-            questions_str = await interrogator_chain.ainvoke(
-                {
-                    "original_request": state["original_request"],
-                    "num_questions": num_questions,
-                    "further_questions": user_questions,
-                }
-            )
-            questions_data = clean_and_parse_json(questions_str)
-            questions = questions_data.get("questions", [])
-            if not questions:
-                raise ValueError("No questions generated by interrogator.")
-            await log_stream.put(
-                f"SUCCESS: Generated {len(questions)} expert questions."
-            )
-        except Exception as e:
-            await log_stream.put(
-                f"ERROR: Failed to generate questions for harvesting. Error: {e}. Aborting harvest."
-            )
-            return {"academic_papers": {}}
-
-        paper_formatter_chain = get_paper_formatter_chain(formatter_llm)
-        academic_papers = {}
-
-        MAX_CONTEXT_CHARS = 250000
-
-        generation_tasks = []
-
-        for question in questions:
-
-            async def generate_paper(q):
-                try:
-                    await log_stream.put(
-                        f"LOG: [HARVEST] Processing Question: '{q[:100]}...'"
-                    )
-                    retrieved_docs = raptor_index.retrieve(q, k=40)
-
-                    if not retrieved_docs:
-                        await log_stream.put(
-                            f"WARNING: No relevant documents found for question '{q[:50]}...'. Skipping paper generation."
-                        )
-                        return None, None
-
-                    await log_stream.put(
-                        f"LOG: Retrieved {len(retrieved_docs)} documents from RAG index for question."
-                    )
-                    rag_context = "\n\n---\n\n".join(
-                        [doc.page_content for doc in retrieved_docs]
-                    )
-
-                    if len(rag_context) > MAX_CONTEXT_CHARS:
-                        await log_stream.put(
-                            f"WARNING: RAG context length ({len(rag_context)} chars) exceeds limit. Truncating to {MAX_CONTEXT_CHARS} chars."
-                        )
-                        rag_context = rag_context[:MAX_CONTEXT_CHARS]
-
-                    paper_content = await paper_formatter_chain.ainvoke(
-                        {"question": q, "rag_context": rag_context}
-                    )
-                    await log_stream.put(
-                        f"SUCCESS: Generated document for question '{q[:50]}...'."
-                    )
-                    return q, paper_content
-                except Exception as e:
-                    await log_stream.put(
-                        f"ERROR: Failed during document generation for question '{q[:50]}...'. Error: {e}"
-                    )
-                    return None, None
-
-            generation_tasks.append(generate_paper(question))
-
-        results = await asyncio.gather(*generation_tasks)
-        for question, paper_content in results:
-            if question and paper_content:
-                academic_papers[question] = paper_content
-
-        await log_stream.put(
-            f"--- [FINAL HARVEST] Finished. Generated {len(academic_papers)} papers. ---"
-        )
-        return {"academic_papers": academic_papers}
-
-    return final_harvest_node
-
-
-@app.get("/", response_class=HTMLResponse)
+@app.get("/")
 def get_index():
     with open("index.html", "r", encoding="utf-8") as f:
         return HTMLResponse(content=f.read(), status_code=200)
@@ -2936,6 +665,11 @@ async def build_and_run_graph(payload: dict = Body(...)):
             "  Phase 0 Foundation → Phase 1 Grid → Phase 2 Noise\n"
             "  → Phase 3 Denoise → Phase 4 Synthesis"
         )
+        qdad_est = estimate_qdad_cost(
+            n=int(params.get("grid_size", params.get("n", 3)) or 3),
+            denoising_steps=int(params.get("denoising_steps", 2) or 2),
+        )
+        await log_stream.put(f"LOG: [COST] {qdad_est.summary_line()}")
         asyncio.create_task(
             run_qdad_background(
                 llm=llm,
@@ -2956,7 +690,11 @@ async def build_and_run_graph(payload: dict = Body(...)):
             )
         )
         return JSONResponse(
-            content={"message": "Graph started.", "session_id": session_id}
+            content={
+                "message": "Graph started.",
+                "session_id": session_id,
+                "cost_estimate": qdad_est.to_dict(),
+            }
         )
 
     # ── Brainstorm (QNN): same library engine as `deepthink qnn` / /qnn skill ──
@@ -2985,6 +723,14 @@ async def build_and_run_graph(payload: dict = Body(...)):
             "  → Epochs (forward / map / Mirror Descent / reframe)\n"
             "  → Solution-Space Report"
         )
+        qnn_est = estimate_qnn_cost(
+            layers=int(params.get("manual_layers", 3) or 3),
+            width=int(params.get("manual_width", 3) or 3),
+            epochs=int(params.get("num_epochs", 2) or 2),
+            qnn_mode=str(params.get("qnn_mode", "auto")),
+            has_document_context=bool(document_context),
+        )
+        await log_stream.put(f"LOG: [COST] {qnn_est.summary_line()}")
         asyncio.create_task(
             run_qnn_background(
                 llm=llm,
@@ -2997,576 +743,23 @@ async def build_and_run_graph(payload: dict = Body(...)):
             )
         )
         return JSONResponse(
-            content={"message": "Graph started.", "session_id": session_id}
+            content={
+                "message": "Graph started.",
+                "session_id": session_id,
+                "cost_estimate": qnn_est.to_dict(),
+            }
         )
 
-    decomposed_problems_map = {}
-    all_layers_prompts = []
-    agent_personas = {}
 
-    try:
-        if mode == "brainstorm":
-            try:
-                # BRAINSTORM = full QNN expert panel (same algorithm as portable /qnn skill)
-                await log_stream.put(
-                    "--- [QNN] Brainstorm expert panel: deepthink algorithm setup ---"
-                )
-                await log_stream.put(
-                    "LOG: [QNN] Steps: 0 Brief → 1 Topology → 2 Seeds → 3 Personas → "
-                    "4 Epochs (forward / map / Mirror Descent / reframe) → 5 Solution-Space Report"
-                )
-
-                # Extract chat history and document context from payload
-                chat_history = payload.get("chat_history", [])
-                document_context = payload.get("document_context", "")
-
-                # Format chat history as string for context
-                chat_history_str = ""
-                if chat_history:
-                    chat_history_str = "\n".join(
-                        [
-                            f"{'User' if msg.get('role') == 'user' else 'Assistant'}: {msg.get('content', '')}"
-                            for msg in chat_history
-                        ]
-                    )
-
-                if document_context:
-                    await log_stream.put(
-                        f"LOG: Document context provided ({len(document_context)} characters)."
-                    )
-
-                # --- QNN Step 0: Impasse / Enrich Brief ---
-                await log_stream.put("LOG: [QNN STEP 0] Building Impasse/Enrich brief...")
-                brainstorm_problem_summary = ""
-                if document_context:
-                    summarizer_chain = get_problem_summarizer_chain(llm)
-                    brainstorm_problem_summary = await summarizer_chain.ainvoke(
-                        {
-                            "user_input": user_prompt,
-                            "document_context": document_context[:50000],
-                        }
-                    )
-                else:
-                    # Still produce a compact brief from the prompt alone
-                    brainstorm_problem_summary = user_prompt
-                await log_stream.put(
-                    f"LOG: [QNN STEP 0] Brief ready ({len(brainstorm_problem_summary)} chars)."
-                )
-
-                # QNN Topology: manual/massive or auto estimator
-                qnn_mode = params.get("qnn_mode", "auto")
-                if "num_epochs" not in params:
-                    params["num_epochs"] = 2
-                else:
-                    params["num_epochs"] = int(params.get("num_epochs", 2))
-
-                width = 3
-                cot_trace_depth = 2
-
-                if qnn_mode == "manual":
-                    try:
-                        raw_layers = int(params.get("manual_layers", 5))
-                        raw_width = int(params.get("manual_width", 5))
-                        cot_trace_depth = max(1, raw_layers)
-                        width = max(1, raw_width)
-                        await log_stream.put(
-                            "--- [QNN STEP 1] USER MANUAL / MASSIVE TOPOLOGY ---"
-                        )
-                        await log_stream.put(
-                            f"LOG: [QNN STEP 1] Manual topology: {cot_trace_depth}L × {width}W × "
-                            f"{params['num_epochs']}E "
-                            f"({cot_trace_depth * width} agents). Can be huge."
-                        )
-                    except Exception as e:
-                        await log_stream.put(
-                            f"WARNING: Bad manual_layers/width from UI, using small defaults. Error: {e}"
-                        )
-                        cot_trace_depth = 5
-                        width = 5
-                else:
-                    # AUTO: complexity estimator picks L, W, E (UI epochs overridden by estimator)
-                    await log_stream.put(
-                        "--- [QNN STEP 1] Auto topology via complexity estimator ---"
-                    )
-
-                    complexity_chain = get_complexity_estimator_chain(llm)
-                    complexity_result_str = await complexity_chain.ainvoke(
-                        {
-                            "user_input": user_prompt,
-                            "prior_conversation": chat_history_str,
-                            "document_context": document_context[:10000]
-                            if document_context
-                            else "",
-                        }
-                    )
-
-                    try:
-                        complexity_data = clean_and_parse_json(complexity_result_str)
-                        if complexity_data is None:
-                            raise ValueError(
-                                "Failed to parse complexity estimation response"
-                            )
-
-                        cot_trace_depth = max(
-                            1, int(complexity_data.get("recommended_layers", 2))
-                        )
-                        width = max(
-                            1, int(complexity_data.get("recommended_width", 3))
-                        )
-                        # Auto mode uses estimator epochs (skill-aligned)
-                        params["num_epochs"] = max(
-                            1, int(complexity_data.get("recommended_epochs", 2))
-                        )
-                        score = complexity_data.get("complexity_score", "?")
-                        reason = complexity_data.get("reasoning", "")
-
-                        await log_stream.put(
-                            f"LOG: [QNN STEP 1] Auto topology: {cot_trace_depth}L × {width}W × "
-                            f"{params['num_epochs']}E "
-                            f"(score={score}, agents={cot_trace_depth * width}). {reason}"
-                        )
-                    except Exception as e:
-                        await log_stream.put(
-                            f"WARNING: Complexity estimation failed. Using defaults. Error: {e}"
-                        )
-                        cot_trace_depth = 2
-                        width = 3
-                        params["num_epochs"] = max(1, int(params.get("num_epochs", 2)))
-
-                # --- QNN Step 2: Seed verbs + nouns (same spanning method as Algorithm Mode) ---
-                # Algorithm: seed_generation → pool of verbs → sample vector_word_size per column.
-                # Brainstorm: seed verbs AND nouns from problem space (+ far fields) → sample per column.
-                vector_word_size = max(
-                    2, int(params.get("vector_word_size", 6))
-                )
-                total_seed_words = max(
-                    vector_word_size * width, vector_word_size * 2
-                )
-                await log_stream.put(
-                    f"LOG: [QNN STEP 2] Generating {total_seed_words} seed verbs+nouns "
-                    f"(vector_word_size={vector_word_size}, width={width})..."
-                )
-                seed_chain = get_brainstorming_seed_chain(llm)
-                seeds_str = await seed_chain.ainvoke(
-                    {"problem": user_prompt, "word_count": total_seed_words}
-                )
-                all_seed_words = list(
-                    {
-                        w.strip()
-                        for w in seeds_str.replace(",", " ").split()
-                        if w.strip() and len(w.strip()) > 1
-                    }
-                )
-                # Fallback fillers if the LLM under-produces
-                fallback_words = [
-                    "distill",
-                    "reconverge",
-                    "entangle",
-                    "ownership",
-                    "latch",
-                    "invariant",
-                    "horizon",
-                    "entropy",
-                    "braid",
-                    "crystallize",
-                    "probe",
-                    "reframe",
-                ]
-                while len(all_seed_words) < total_seed_words:
-                    all_seed_words.append(
-                        fallback_words[len(all_seed_words) % len(fallback_words)]
-                    )
-                random.shuffle(all_seed_words)
-
-                # One guiding word-vector per column (shared across layers, like algorithm MBTI seeds)
-                column_guiding_words = []
-                for j in range(width):
-                    if len(all_seed_words) >= vector_word_size:
-                        sample = random.sample(all_seed_words, vector_word_size)
-                    else:
-                        sample = list(all_seed_words)
-                    column_guiding_words.append(" ".join(sample))
-
-                await log_stream.put(
-                    f"LOG: [QNN STEP 2] Seed pool ({len(all_seed_words)} words): "
-                    f"{' '.join(all_seed_words[: min(24, len(all_seed_words))])}"
-                    f"{'...' if len(all_seed_words) > 24 else ''}"
-                )
-                for j, gw in enumerate(column_guiding_words):
-                    await log_stream.put(
-                        f"LOG: [QNN STEP 2] Column {j} guiding_words: {gw}"
-                    )
-
-                # --- QNN Step 3: Span L×W personas from guiding_words (input-spanner style) ---
-                await log_stream.put(
-                    f"LOG: [QNN STEP 3] Spanning {cot_trace_depth}×{width} personas "
-                    f"from verb/noun word-vectors..."
-                )
-                spanner_chain = get_brainstorming_spanner_chain(llm)
-
-                for i in range(cot_trace_depth):
-                    layer_role = "DIVERGENT" if i == 0 else "CONVERGENT"
-                    layer_prompts = []
-                    layer_tasks = []
-                    for j in range(width):
-                        layer_tasks.append(
-                            spanner_chain.ainvoke(
-                                {
-                                    "problem": user_prompt,
-                                    "guiding_words": column_guiding_words[j],
-                                    "layer_index": i,
-                                    "node_index": j,
-                                    "document_context": brainstorm_problem_summary,
-                                }
-                            )
-                        )
-
-                    personas_raw = await asyncio.gather(*layer_tasks)
-
-                    for j, p_str in enumerate(personas_raw):
-                        agent_id = f"agent_{i}_{j}"
-                        gw = column_guiding_words[j]
-                        try:
-                            persona = clean_and_parse_json(p_str)
-                            if not isinstance(persona, dict):
-                                raise ValueError("persona not a dict")
-                        except Exception:
-                            persona = {
-                                "name": f"Expert {i}-{j}",
-                                "specialty": f"Specialist shaped by {gw}",
-                                "emoji": "🧠",
-                                "guiding_words": gw,
-                                "attributes": gw.split(),
-                                "skills": ["problem-space spanning", "strategic mapping"],
-                                "system_prompt": (
-                                    f"You are a {layer_role} QNN node. Your cognition is shaped by "
-                                    f"these guiding words: {gw}. Specialize for: {user_prompt}. "
-                                    "Map strategies with mechanisms and falsifiers; no production patches."
-                                ),
-                            }
-
-                        attrs = persona.get("attributes") or []
-                        if isinstance(attrs, list):
-                            attrs_block = "\n".join(f"- {a}" for a in attrs[:12])
-                        else:
-                            attrs_block = str(attrs)
-                        skills = persona.get("skills") or []
-                        if isinstance(skills, list):
-                            skills_block = "\n".join(f"* {s}" for s in skills[:6])
-                        else:
-                            skills_block = str(skills)
-
-                        system_prompt = f"""
-You are {persona.get("name", "Expert")} {persona.get("emoji", "🧠")}.
-Your Specialty is: {persona.get("specialty", "Analysis")}.
-QNN cell: Layer {i} ({layer_role}), Node {j}.
-
-### Guiding Words (problem-space verb/noun vector)
-{gw}
-
-### Attributes
-{attrs_block or "- (derived from guiding words)"}
-
-### Skills
-{skills_block or "* strategic mapping"}
-
-<Role>
-{persona.get("system_prompt", "Analyze the input through your guiding words.")}
-</Role>
-
-<QNN Discipline>
-- You are a neuron in a layered Qualitative Neural Network, not a flat chat expert.
-- Your persona was spanned from problem-space verbs and nouns (same method as Algorithm Mode).
-- Layer 0: diverge. Deeper layers: critique/refine upstream (cite agent ids).
-- Do NOT write production patches or full file diffs.
-- Produce strategy angles with mechanisms, falsifiers, and risks.
-</QNN Discipline>
-"""
-                        layer_prompts.append(system_prompt)
-
-                        agent_personas[agent_id] = {
-                            "name": persona.get("name", f"Agent {i}-{j}"),
-                            "mbti_type": "Expert",
-                            "specialty": persona.get("specialty", "Analysis"),
-                            "guiding_words": gw,
-                            "layer": i,
-                            "layer_role": layer_role,
-                        }
-                        decomposed_problems_map[agent_id] = user_prompt
-
-                    all_layers_prompts.append(layer_prompts)
-                    await log_stream.put(
-                        f"LOG: [QNN STEP 3] Layer {i} ({layer_role}): {width} personas spanned."
-                    )
-
-                await log_stream.put(
-                    f"LOG: [QNN STEP 3] Expert panel ready: {cot_trace_depth}L × {width}W = "
-                    f"{cot_trace_depth * width} agents; epochs={params['num_epochs']}."
-                )
-                await log_stream.put(
-                    "LOG: [QNN STEP 4] Epoch loop starts with graph execution "
-                    "(layered forward → map → Mirror Descent → reframe)..."
-                )
-
-            except Exception as e:
-                await log_stream.put(
-                    f"ERROR: Error during brainstorming graph setup: {e}"
-                )
-                await log_stream.put(traceback.format_exc())
-                # Fallback to a minimal setup so the graph can at least run
-                await log_stream.put(
-                    "LOG: Falling back to minimal brainstorming topology."
-                )
-                params["num_epochs"] = 1
-                width = 3
-                guiding_concepts = ["General_Analysis"] * 3
-                all_layers_prompts = [
-                    [
-                        f"You are a generic brainstorming agent. Analyze the topic: {user_prompt}"
-                    ]
-                    * 3
-                ]
-                for j in range(3):
-                    agent_id = f"agent_0_{j}"
-                    agent_personas[agent_id] = {
-                        "name": f"Expert {j}",
-                        "mbti_type": "Expert",
-                        "specialty": "Generalist",
-                    }
-                    decomposed_problems_map[agent_id] = user_prompt
-
-        else:
-            return JSONResponse(
-                content={
-                    "message": (
-                        f"Unsupported mode '{mode}'. "
-                        "Use 'brainstorm' or 'app_slot_machine'."
-                    )
-                },
-                status_code=400,
+    return JSONResponse(
+        content={
+            "message": (
+                f"Unsupported mode '{mode}'. "
+                "Use 'brainstorm' or 'app_slot_machine'."
             )
-
-    except Exception as e:
-        error_message = f"Error during graph setup: {e}"
-        await log_stream.put(error_message)
-        await log_stream.put(traceback.format_exc())
-        return JSONResponse(content={"message": error_message}, status_code=500)
-
-    # Building Graph Nodes
-    workflow = StateGraph(GraphState)
-
-    # Per-agent model support: cycle through agent_model_list (or default)
-    # Build llm per agent node
-    effective_models = agent_model_list if agent_model_list else [default_agent_model]
-    for i, layer_prompts in enumerate(all_layers_prompts):
-        for j, _ in enumerate(layer_prompts):
-            node_id = f"agent_{i}_{j}"
-            m_idx = (i * max(1, len(layer_prompts)) + j) % len(effective_models)
-            model_for_this_agent = effective_models[m_idx]
-            if is_debug:
-                per_agent_llm = CoderMockLLM()
-            elif provider == "openrouter":
-                try:
-                    per_agent_llm = ChatOpenAI(
-                        model=model_for_this_agent,
-                        openai_api_key=api_key,
-                        openai_api_base="https://openrouter.ai/api/v1",
-                        temperature=0.7,
-                        callbacks=[token_tracker],
-                    )
-                except Exception:
-                    per_agent_llm = llm
-            else:
-                try:
-                    per_agent_llm = ChatLlamaCpp(
-                        base_url=llamacpp_url,
-                        api_key=llamacpp_api_key,
-                        model=model_for_this_agent,
-                        temperature=0.7,
-                        max_tokens=4096,
-                    )
-                except Exception:
-                    per_agent_llm = llm
-            workflow.add_node(node_id, create_agent_node(per_agent_llm, node_id))
-
-    workflow.add_node("synthesis", create_synthesis_node(synthesis_llm))
-    workflow.add_node("code_execution", create_code_execution_node(llm))
-    workflow.add_node("archive_epoch", create_archive_epoch_outputs_node())
-    workflow.add_node(
-        "update_rag_index", create_update_rag_index_node(llm, embeddings_model)
-    )  # Added RAG node
-    workflow.add_node("metrics", create_metrics_node(llm))
-    workflow.add_node("reframe_and_decompose", create_reframe_and_decompose_node(llm))
-    workflow.add_node("update_prompts", create_update_agent_prompts_node(llm))
-
-    # Add Edges (Architecture)
-    # Layer 0 -> Layer 1 ... -> Synthesis
-
-    first_layer_nodes = [f"agent_0_{j}" for j in range(len(all_layers_prompts[0]))]
-
-    # Parallel Entry: Connect START to ALL Layer 0 nodes
-    for n in first_layer_nodes:
-        workflow.add_edge(START, n)
-
-    for i in range(len(all_layers_prompts) - 1):
-        current_layer_nodes = [
-            f"agent_{i}_{j}" for j in range(len(all_layers_prompts[i]))
-        ]
-        next_layer_nodes = [
-            f"agent_{i + 1}_{k}" for k in range(len(all_layers_prompts[i + 1]))
-        ]
-        for curr in current_layer_nodes:
-            for nxt in next_layer_nodes:
-                workflow.add_edge(curr, nxt)
-
-    last_layer_nodes = [
-        f"agent_{len(all_layers_prompts) - 1}_{j}"
-        for j in range(len(all_layers_prompts[-1]))
-    ]
-    for n in last_layer_nodes:
-        workflow.add_edge(n, "synthesis")
-
-    workflow.add_edge("synthesis", "code_execution")
-    workflow.add_edge("code_execution", "archive_epoch")
-    workflow.add_edge("archive_epoch", "update_rag_index")  # Route to RAG index update
-    workflow.add_edge("update_rag_index", "metrics")  # Route to metrics after indexing
-
-    # Conditional Edge for Loops
-    def epoch_gateway(state):
-        # We start at epoch 0. We want a total of 'max_epochs' passes.
-        # Pass 0 ends here. If max_epochs is 2, we want Pass 0 and Pass 1.
-        # So at Pass 0 (epoch=0), we check: 0 < 2-1 (0 < 1) -> True -> Loop.
-        # At Pass 1 (epoch=1), we check: 1 < 2-1 (1 < 1) -> False -> Stop.
-        if state["epoch"] < state["max_epochs"] - 1:
-            return "reframe_and_decompose"
-        return "harvest"
-
-    workflow.add_conditional_edges(
-        "metrics",
-        epoch_gateway,
-        {"reframe_and_decompose": "reframe_and_decompose", "harvest": END},
+        },
+        status_code=400,
     )
-    # Note: Using END here effectively means we break the loop if done.
-    # But wait, we need to add Harvest Node to graph if we link to it?
-    # Or handled by app logic?
-    # Original logic had "harvest" key mapping to... END?
-    # Actually, harvest is usually a separate call or node.
-    # We should map "harvest" to END, and let the frontend call /harvest if needed?
-    # OR create a harvest node?
-    # The original code had conditional edge to "reframe..." or END.
-
-    workflow.add_edge("reframe_and_decompose", "update_prompts")
-    # Loop back to Entry Point is tricky with LangGraph.
-    # We need to restart the agent nodes.
-    # Connect update_prompts to Layer 0 nodes?
-    for n in first_layer_nodes:
-        workflow.add_edge("update_prompts", n)
-
-    graph = workflow.compile()
-
-    ascii_art = graph.get_graph().draw_ascii()
-    await log_stream.put(ascii_art)
-
-    session_id = str(uuid.uuid4())
-
-    # Prepare brainstorm context (only relevant in brainstorm mode, but always include empty defaults)
-    brainstorm_chat_history_str = ""
-    brainstorm_document_context = ""
-    brainstorm_problem_summary_state = ""
-    if mode == "brainstorm":
-        brainstorm_chat_history_str = chat_history_str
-        brainstorm_document_context = document_context
-        # Set during QNN Step 0 setup; fallback to prompt if setup path skipped it
-        brainstorm_problem_summary_state = (
-            locals().get("brainstorm_problem_summary") or user_prompt
-        )
-
-    initial_state = {
-        "session_id": session_id,
-        "mode": mode,
-        "original_request": user_prompt,
-        "current_problem": user_prompt,
-        "decomposed_problems": decomposed_problems_map,
-        "epoch": 0,
-        "max_epochs": int(params.get("num_epochs", 1)),
-        "params": params,
-        "all_layers_prompts": all_layers_prompts,
-        "agent_personas": agent_personas,
-        "is_code_request": is_code,
-        "agent_outputs": {},
-        "memory": {},
-        "final_solution": None,
-        "previous_solution": "",
-        "chat_history": [],
-        "layers": [],
-        "critiques": {},
-        "perplexity_history": [],
-        "raptor_index": None,
-        "all_rag_documents": [],
-        "academic_papers": None,
-        "summarizer_llm": summarizer_llm,
-        "embeddings_model": embeddings_model,
-        "modules": [],
-        "synthesis_context_queue": [],
-        "synthesis_execution_success": True,
-        # Brainstorm / QNN mode context
-        "brainstorm_prior_conversation": brainstorm_chat_history_str,
-        "brainstorm_document_context": brainstorm_document_context,
-        "brainstorm_problem_summary": brainstorm_problem_summary_state,
-        # Qualitative Self-Attention edges (brainstorm): filled during forward pass
-        "attention_edges": {},
-    }
-    initial_state["llm"] = llm
-    sessions[session_id] = initial_state
-
-    await log_stream.put(f"__session_id__ {session_id}")
-    await log_stream.put(f"__start__ {ascii_art}")  # Send start signal + ASCII
-
-    # Run Graph
-    asyncio.create_task(run_graph_background(graph, initial_state))
-
-    return JSONResponse(content={"message": "Graph started.", "session_id": session_id})
-
-
-async def run_graph_background(graph, initial_state):
-    session_id = initial_state["session_id"]
-    mode = initial_state.get("mode", "algorithm")
-    try:
-        async for output in graph.astream(initial_state, {"recursion_limit": 100}):
-            for node_name, node_output in output.items():
-                if node_output:
-                    # Update session state
-                    current = sessions[session_id]
-                    for k, v in node_output.items():
-                        if isinstance(current.get(k), dict) and isinstance(v, dict):
-                            current[k].update(v)
-                        elif isinstance(current.get(k), list) and isinstance(v, list):
-                            current[k].extend(v)
-                        else:
-                            current[k] = v
-                    sessions[session_id] = current
-
-        # Final result processing after the graph ends
-        final_state = sessions[session_id]
-        if mode == "brainstorm":
-            # Brainstorm mode already emits FINAL_ANSWER from the synthesis node itself,
-            # but we log completion here for the server console.
-            await log_stream.put("SUCCESS: Brainstorm graph execution completed.")
-        elif mode == "app_slot_machine":
-            await log_stream.put("SUCCESS: App Slot Machine (QDAD) completed.")
-        else:
-            final_solution = final_state.get("final_solution")
-            if final_solution:
-                await log_stream.put(f"FINAL_ANSWER: {json.dumps(final_solution)}")
-                await log_stream.put("SUCCESS: Graph execution completed.")
-            else:
-                await log_stream.put(
-                    "WARNING: Graph completed but no final_solution was found in state."
-                )
-
-    except Exception as e:
-        await log_stream.put(f"Graph Background Error: {e}")
-        await log_stream.put(traceback.format_exc())
 
 
 @app.get("/export_qnn/{session_id}")
