+++ deepthink/nodes/__init__.py
@@ -1,1 +1,13 @@
-# nodes subpackage
+"""Compatibility shim — node factories live in ``deepthink.runtime.nodes``."""
+
+from deepthink.runtime.nodes import (  # noqa: F401
+    create_agent_node,
+    create_archive_epoch_outputs_node,
+    create_code_execution_node,
+    create_final_harvest_node,
+    create_metrics_node,
+    create_reframe_and_decompose_node,
+    create_synthesis_node,
+    create_update_agent_prompts_node,
+    create_update_rag_index_node,
+)
