+++ deepthink/qnn/__init__.py
@@ -8,10 +8,16 @@ Portable entrypoint for skills and harnesses:
 See ``run_qnn_pipeline`` for the full parameter surface.
 """
 
-from .pipeline import run_qnn_pipeline, QNNResult, default_qnn_params
+from .pipeline import (
+    QNNResult,
+    clamp_qnn_topology,
+    default_qnn_params,
+    run_qnn_pipeline,
+)
 
 __all__ = [
     "run_qnn_pipeline",
     "QNNResult",
     "default_qnn_params",
+    "clamp_qnn_topology",
 ]
