+++ skills/qdad/run_template.py
@@ -148,9 +148,9 @@ if __name__ == "__main__":
     p.add_argument("--api-key", default=None)
     p.add_argument("--model", default="stepfun/step-3.5-flash:free")
     p.add_argument("--base-url", default="http://localhost:8080/v1")
-    p.add_argument("--n", type=int, default=4)
+    p.add_argument("--n", type=int, default=3)
     p.add_argument("--temperature-scale", type=float, default=1.3)
-    p.add_argument("--denoising-steps", type=int, default=3)
+    p.add_argument("--denoising-steps", type=int, default=2)
     p.add_argument("--noun-verb-temperature", type=float, default=0.6)
     p.add_argument("--context-file", default=None)
     p.add_argument("--out", default=None)
